Module Beluga

module Abstract : sig ... end

module Apxnorm : sig ... end

module Chatter : sig ... end

module Check : sig ... end

module Command : sig ... end

module Context : sig ... end

Contexts

module ConvSigma : sig ... end

module Coverage : sig ... end

module Ctxsub : sig ... end

module Erase : sig ... end

module Holes : sig ... end

module Index : sig ... end

Name resolution and translation from the concrete syntax to the approximate syntax.

module Index_state : sig ... end

State definition for the indexing phase.

module Interactive : sig ... end

module Lfcheck : sig ... end

module Lfrecon : sig ... end

module Load : sig ... end

module Logic : sig ... end

module Monitor : sig ... end

Timing utilities for monitoring the execution time of functions.

module NameGen : sig ... end

Contextual type-based name generator. This name generator will use programmer-supplied naming hints (falling back to good old X, Y, Z) but adjust the numeric counter on the name so it is the smallest allowed in a given context.

module Opsem : sig ... end

module Options : sig ... end

Centralized module for global flags.

module Order : sig ... end

module Prettyint : sig ... end

module Printer : sig ... end

module Reconstruct : sig ... end

module Recsgn : sig ... end

module Recsgn_state : sig ... end

module Sigma : sig ... end

module Store : sig ... end

module Subord : sig ... end

module Substitution : sig ... end

Substitutions

module Total : sig ... end

Checking termination of a function

module Trail : sig ... end

Trailing Abstract Operations

module Typeinfo : sig ... end

module Unify : sig ... end

module Whnf : sig ... end