Module Check.Comp

type typeVariant =

| VariantCross

| VariantArrow

| VariantCtxPi

| VariantPiBox

| VariantBox

type mismatch_kind = [

| `fn

| `mlam

| `box

| `ctxfun

| `tuple

]

type error =

| IllegalParamTyp of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.typ

| MismatchChk of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.Comp.gctx * Beluga_syntax.Synint.Comp.exp * Beluga_syntax.Synint.Comp.tclo * Beluga_syntax.Synint.Comp.tclo

| MismatchSyn of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.Comp.gctx * Beluga_syntax.Synint.Comp.exp * typeVariant * Beluga_syntax.Synint.Comp.tclo

| PatIllTyped of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.Comp.gctx * Beluga_syntax.Synint.Comp.pattern * Beluga_syntax.Synint.Comp.tclo * Beluga_syntax.Synint.Comp.tclo

| BasicMismatch of mismatch_kind * Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.Comp.gctx * Beluga_syntax.Synint.Comp.tclo

| SBoxMismatch of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.Comp.gctx * Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.dctx

| SynMismatch of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.Comp.tclo * Beluga_syntax.Synint.Comp.tclo

| TypMismatch of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.Comp.tclo * Beluga_syntax.Synint.Comp.tclo

| UnsolvableConstraints of Beluga_syntax.Name.t option * string

| InvalidRecCall

| NotRecursiveSrc of Beluga_syntax.Name.t

| NotRecursiveDst of Beluga_syntax.Name.t

| MissingTotal of Beluga_syntax.Id.cid_prog

| NotImpossible of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.Comp.gctx * Beluga_syntax.Synint.Comp.typ * Beluga_syntax.Synint.Comp.exp

| InvalidHypotheses of Beluga_syntax.Synint.Comp.hypotheses * Beluga_syntax.Synint.Comp.hypotheses

| SufficesDecompositionFailed of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.Comp.typ

| SufficesLengthsMismatch of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.Comp.typ * int * int

| SufficesBadAnnotation of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.Comp.typ * int * Beluga_syntax.Synint.Comp.typ

| SufficesBadGoal of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.Comp.typ * Beluga_syntax.Synint.Comp.typ

| SufficesPremiseNotClosed of Beluga_syntax.Synint.LF.mctx * int * Beluga_syntax.Synint.Comp.suffices_typ

exception Error of Beluga_syntax.Location.t * error

val throw : Beluga_syntax.Location.t -> error -> 'a

Raises an error from this module.

val append_hypotheses : Beluga_syntax.Synint.Comp.hypotheses -> Beluga_syntax.Synint.Comp.hypotheses -> Beluga_syntax.Synint.Comp.hypotheses

Appends two sets of hypotheses. Appropriately MShifts the left contexts and applies an appropriate totality shift.

val apply_unbox_modifier : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.Comp.unbox_modifier -> Beluga_syntax.Synint.LF.ctyp -> Beluga_syntax.Synint.LF.ctyp * Beluga_syntax.Synint.LF.sub

Transforms the given contextual type according the an unboxing modifier. The returned substitution witnesses the transformation of the type. It should be applied to the metavariable of the computed type in order for it to make sense in the original context. For example, suppose apply_unbox_modifier cD `Strengthen cU = cU', s If we want X : cU' to make sense in its original LF context, then it suffices to apply Xs. This substitution is used by the translation from proofs into programs.

val apply_unbox_modifier_opt : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.Comp.unbox_modifier option -> Beluga_syntax.Synint.LF.ctyp -> Beluga_syntax.Synint.LF.ctyp * Beluga_syntax.Synint.LF.sub

Variant of apply_unbox_modifier that is the identity when no modifier is specified.

val validate_contexts : Beluga_syntax.Location.t -> (Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.LF.mctx) -> (Beluga_syntax.Synint.Comp.gctx * Beluga_syntax.Synint.Comp.gctx) -> Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.Comp.gctx

Verifies that the pairs of contexts are convertible. It is a user error InvalidHypotheses if they are not. If the contexts are convertible, then they are merged, using plicity and type information from the context on the left, but the entry name from the context on the right.

val thm : Beluga_syntax.Id.cid_prog option -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.Comp.gctx -> Beluga_syntax.Synint.Comp.total_dec list -> ?cIH:Beluga_syntax.Synint.Comp.ihctx -> Beluga_syntax.Synint.Comp.thm -> Beluga_syntax.Synint.Comp.tclo -> unit

Checks a theorem in the given contexts against the given type. The given list of total declarations is used for totality checking. The cid_comp_const parameter is used for registering Harpoon subgoals. For ordinary Beluga programs or for complete Harpoon proofs, this argument can be None.

val check : Beluga_syntax.Id.cid_prog option -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.Comp.gctx -> Beluga_syntax.Synint.Comp.total_dec list -> ?cIH:Beluga_syntax.Synint.Comp.ihctx -> Beluga_syntax.Synint.Comp.exp -> Beluga_syntax.Synint.Comp.tclo -> unit

val syn : Beluga_syntax.Id.cid_prog option -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.Comp.gctx -> Beluga_syntax.Synint.Comp.total_dec list -> ?cIH:Beluga_syntax.Synint.Comp.ihctx -> Beluga_syntax.Synint.Comp.exp -> Beluga_syntax.Synint.Comp.ihctx option * Beluga_syntax.Synint.Comp.tclo

val checkKind : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.Comp.kind -> unit

val checkTyp : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.Comp.typ -> unit

val wf_mctx : Beluga_syntax.Synint.LF.mctx -> unit

val mvars_in_patt : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.Comp.pattern -> Beluga_syntax.Synint.LF.mctx

Transforms the given meta-context by marking all meta-variables appearing in the pattern as Inductive.

This function relies on numerous (non-exported; 87 lines total) helpers to traverse the syntax tree. I believe that there is an opportunity to refactor this and move it to a different module. -je

val fixParamTyp : Beluga_syntax.Synint.LF.mfront -> Beluga_syntax.Synint.Comp.meta_typ -> int option * Beluga_syntax.Synint.Comp.meta_typ * int option

Analyzes a contextual object to decide whether it's a (contextual) variable and rewrites its type from MTyp to PTyp if necessary. This is crucial for coverage checking of a case analysis on a parameter variable. Also returns the index of the variable (so we can later decide if it's a variable we're doing induction on) as well what projection, if any, is applied to the parameter variable.

val id_map_ind : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.msub -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.mctx

Transfers inductivity annotations from a source context to a target context related by a meta-substitution.

id_map_ind cD' t cD = cD'* where cD' |- t : cD

val unroll : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.Comp.gctx -> Beluga_syntax.Synint.Comp.ihctx -> Beluga_syntax.Synint.Comp.typ -> Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.Comp.gctx * Beluga_syntax.Synint.Comp.ihctx * Beluga_syntax.Synint.Comp.typ * Beluga_syntax.Synint.LF.msub

unroll cD cG cIH tau = (cD', cG', cIH', tau', t) where cD |- cG <= ctx and cD |- tau <= type s.t. cD' extends cD and cG' extends cG and tau' is a subterm of tau and tau' is not an arrow nor a PiBox type and cD' |- t : cD is a weakening meta-substitution. and cD' |- cIH' ihctx and cIH' is appropriately shifted

val require_syn_typbox : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.Comp.gctx -> Beluga_syntax.Location.t -> Beluga_syntax.Synint.Comp.exp -> Beluga_syntax.Synint.Comp.tclo -> Beluga_syntax.Synint.Comp.meta_typ * Beluga_syntax.Synint.LF.msub

Requires that the given type be a box-type. require_syn_typbox cD cG loc i (tau, t) = (cU, t) if tau = cU; else, raises a synthesis mismatch error for the expression i saying that the type of i is expected to be a box-type.

val decompose_function_type : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.Comp.typ -> (Beluga_syntax.Synint.Comp.typ list * Beluga_syntax.Synint.Comp.typ) option

Processes all leading PiBoxes to replace them with unification variables in the following type, and returns a further decomposition of the type remaining after all these PiBoxes, which must consist only of simple function types (or a base type).

Given cD |- tau <= type and tau = Pi X_1:U_1. ... Pi X_n:U_n. tau' and tau' = tau_1 -> tau_2 -> ... -> tau_k (so cD, X_1:U_1, ..., X_n:U_n |- tau' <= type) `decompose_function_type cD tau` calculates an msub t such that cD |- t : cD, X_1:U_1, ..., X_n:U_n that replaces each X_i with a fresh unification variable ?X_i. The msub t is applied to tau' and tau' is further decomposed into a list l = [t]tau_1; [t]tau_2; ...; [t]tau_(k-1). This list is returned together with the type ttau_k. If the decomposition fails due to an interleaving of PiBox-types and arrow-types, None is returned.

val unify_suffices : Beluga_syntax.Location.t -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.Comp.typ -> Beluga_syntax.Synint.Comp.suffices_typ list -> Beluga_syntax.Synint.Comp.typ -> Beluga_syntax.Synint.Comp.typ list

Checks that the given type annotations are compatible with a given function type.

unify_suffices cD tau_i tau_anns tau_g requires 1. cD |- tau_i <= type 2. for each stau in tau_anns, if stau = `exact tau, then cD |- tau <= type (for each type in the list) 3. cD |- tau_g <= type It decomposes the function type tau_i (see decompose_function_type) and appropriately unifies the decomposition with the given type annotations so as to pin down all instantiations of universally quantified variables. Some type annotations may be omitted. The interpretation of an omitted annotation is that that annotation is not necessary (the type of that argument is already determined by unification of the goal, or other annotations).

Returns the list of elaborated premise types, which are guaranteed to be closed (no MMVars).

The provided location is used when raising errors.

val genMApp : Beluga_syntax.Location.t -> (Beluga_syntax.Synint.LF.ctyp_decl -> bool) -> Beluga_syntax.Synint.LF.mctx -> (Beluga_syntax.Synint.Comp.exp * Beluga_syntax.Synint.Comp.tclo) -> int * (Beluga_syntax.Synint.Comp.exp * Beluga_syntax.Synint.Comp.tclo)

Generates a meta-application spine consisting of unification variables to eliminate leading PiBox types. Returns the number of generated unification variables as well as the type (together with a delayed meta-substitution) of the produced expression.