Module Check.LF

type error =

| CtxVarMisCheck of Beluga_syntax.Int.LF.mctx * Beluga_syntax.Int.LF.dctx * Beluga_syntax.Int.LF.tclo * Beluga_syntax.Name.t * Beluga_syntax.Int.LF.schema

| CtxVarMismatch of Beluga_syntax.Int.LF.mctx * Beluga_syntax.Int.LF.ctx_var * Beluga_syntax.Name.t * Beluga_syntax.Int.LF.schema

| CtxVarDiffer of Beluga_syntax.Int.LF.mctx * Beluga_syntax.Int.LF.ctx_var * Beluga_syntax.Int.LF.ctx_var

| CheckError of Beluga_syntax.Int.LF.mctx * Beluga_syntax.Int.LF.dctx * Beluga_syntax.Int.LF.nclo * Beluga_syntax.Int.LF.tclo

| TupleArity of Beluga_syntax.Int.LF.mctx * Beluga_syntax.Int.LF.dctx * Beluga_syntax.Int.LF.nclo * Beluga_syntax.Int.LF.trec_clo

| SigmaMismatch of Beluga_syntax.Int.LF.mctx * Beluga_syntax.Int.LF.dctx * Beluga_syntax.Int.LF.trec_clo * Beluga_syntax.Int.LF.trec_clo

| KindMismatch of Beluga_syntax.Int.LF.mctx * Beluga_syntax.Int.LF.dctx * Beluga_syntax.Int.LF.sclo * Beluga_syntax.Int.LF.kind * Beluga_syntax.Int.LF.sub

| TypMismatch of Beluga_syntax.Int.LF.mctx * Beluga_syntax.Int.LF.dctx * Beluga_syntax.Int.LF.nclo * Beluga_syntax.Int.LF.tclo * Beluga_syntax.Int.LF.tclo

| IllTypedSub of Beluga_syntax.Int.LF.mctx * Beluga_syntax.Int.LF.dctx * Beluga_syntax.Int.LF.sub * Beluga_syntax.Int.LF.dctx * Beluga_syntax.Int.LF.normal option

| SpineIllTyped of int * int

| LeftoverFV of Beluga_syntax.Name.t

| ParamVarInst of Beluga_syntax.Int.LF.mctx * Beluga_syntax.Int.LF.dctx * Beluga_syntax.Int.LF.tclo

| CtxHatMismatch of Beluga_syntax.Int.LF.mctx * Beluga_syntax.Int.LF.dctx * Beluga_syntax.Int.LF.dctx_hat * Beluga_syntax.Location.t * Beluga_syntax.Int.LF.mfront

| IllTypedMetaObj of Beluga_syntax.Int.LF.mctx * Beluga_syntax.Int.LF.clobj * Beluga_syntax.Int.LF.dctx * Beluga_syntax.Int.LF.cltyp

| TermWhenVar of Beluga_syntax.Int.LF.mctx * Beluga_syntax.Int.LF.dctx * Beluga_syntax.Int.LF.normal

| SubWhenRen of Beluga_syntax.Int.LF.mctx * Beluga_syntax.Int.LF.dctx * Beluga_syntax.Int.LF.sub

| MissingType of string

exception Error of Beluga_syntax.Location.t * error

val throw : Beluga_syntax.Location.t -> error -> 'a

val check : Beluga_syntax.Int.LF.mctx -> Beluga_syntax.Int.LF.dctx -> Beluga_syntax.Int.LF.nclo -> Beluga_syntax.Int.LF.tclo -> unit

val syn : Beluga_syntax.Int.LF.mctx -> Beluga_syntax.Int.LF.dctx -> Beluga_syntax.Int.LF.nclo -> Beluga_syntax.Int.LF.tclo

val checkTyp : Beluga_syntax.Int.LF.mctx -> Beluga_syntax.Int.LF.dctx -> Beluga_syntax.Int.LF.tclo -> unit

val checkKind : Beluga_syntax.Int.LF.mctx -> Beluga_syntax.Int.LF.dctx -> Beluga_syntax.Int.LF.kind -> unit

val checkDCtx : Beluga_syntax.Int.LF.mctx -> Beluga_syntax.Int.LF.dctx -> unit

val checkSchemaWf : Beluga_syntax.Int.LF.schema -> unit

val checkSchema : Beluga_syntax.Location.t -> Beluga_syntax.Int.LF.mctx -> Beluga_syntax.Int.LF.dctx -> Beluga_syntax.Name.t -> Beluga_syntax.Int.LF.schema -> unit

val subsumes : Beluga_syntax.Int.LF.mctx -> Beluga_syntax.Int.LF.ctx_var -> Beluga_syntax.Int.LF.ctx_var -> bool

val checkTypeAgainstSchema : Beluga_syntax.Location.t -> Beluga_syntax.Int.LF.mctx -> Beluga_syntax.Int.LF.dctx -> Beluga_syntax.Int.LF.typ -> Beluga_syntax.Name.t -> Beluga_syntax.Int.LF.sch_elem list -> Beluga_syntax.Int.LF.typ_rec * Beluga_syntax.Int.LF.sub

Checks that a type exists within a given schema. checkTypeAgainstSchema loc cD cPsi tA name es = (tR, s) if cD; cPsi |- tA <== type and tA is an instance of one of the schema elements es. If not, an error will the raised with the given location. The input name will be a part of the error message, and should be the declared name of the schema.

val instanceOfSchElem : Beluga_syntax.Int.LF.mctx -> Beluga_syntax.Int.LF.dctx -> Beluga_syntax.Int.LF.tclo -> Beluga_syntax.Int.LF.sch_elem -> Beluga_syntax.Int.LF.typ_rec * Beluga_syntax.Int.LF.sub

val instanceOfSchElemProj : Beluga_syntax.Int.LF.mctx -> Beluga_syntax.Int.LF.dctx -> Beluga_syntax.Int.LF.tclo -> (Beluga_syntax.Int.LF.head * int) -> Beluga_syntax.Int.LF.sch_elem -> Beluga_syntax.Int.LF.typ_rec * Beluga_syntax.Int.LF.sub

val checkMSub : Beluga_syntax.Location.t -> Beluga_syntax.Int.LF.mctx -> Beluga_syntax.Int.LF.msub -> Beluga_syntax.Int.LF.mctx -> unit