Module Beluga.Command

val register : string -> (Stdlib.Format.formatter -> string list -> unit) -> string -> unit

* Registers a new bespoke command

val is_command : string -> [> `Cmd of string | `Input of string ]

* Decides whether an input string is some Beluga code (constructor * `Input) or an interactive mode command to run (constructor `Cmd).

val do_command : Stdlib.Format.formatter -> string -> unit

Given a formatter for command output, parses the given string as a * command and runs it.

val print_usage : Stdlib.Format.formatter -> unit