Module Beluga.Context

Contexts

author Brigitte Pientka

exception NoTypAvailable

val dec : 'a -> 'a Beluga_syntax.Synint.LF.ctx -> 'a Beluga_syntax.Synint.LF.ctx

Function form of the LF.Dec constructor with the arguments reversed. Useful for extending contexts multiple times without the need for parentheses.

  cD |> Context.dec ... |> Context.dec ...

val decs : 'a list -> 'a Beluga_syntax.Synint.LF.ctx -> 'a Beluga_syntax.Synint.LF.ctx

Extends a context with multiple entries from left to right. That is,

  ctx |> decs [ d1; d2; d3 ]

is equivalent to

  ctx |> dec d1 |> dec d2 |> dec d3

val dctxToHat : Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.dctx_hat

val addToHat : Beluga_syntax.Synint.LF.dctx_hat -> Beluga_syntax.Synint.LF.dctx_hat

val hatToDCtx : Beluga_syntax.Synint.LF.dctx_hat -> Beluga_syntax.Synint.LF.dctx

val extend_hatctx : int -> Beluga_syntax.Synint.LF.dctx_hat -> Beluga_syntax.Synint.LF.dctx_hat

val ctxDec : Beluga_syntax.Synint.LF.dctx -> int -> Beluga_syntax.Synint.LF.typ_decl

val ctxSigmaDec : Beluga_syntax.Synint.LF.dctx -> int -> Beluga_syntax.Synint.LF.typ_decl

val containsSigma : Beluga_syntax.Synint.LF.dctx -> bool

val ctxVar : Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.ctx_var option

val hasCtxVar : Beluga_syntax.Synint.LF.dctx -> bool

val append : 'a Beluga_syntax.Synint.LF.ctx -> 'a Beluga_syntax.Synint.LF.ctx -> 'a Beluga_syntax.Synint.LF.ctx

val fold : 'b -> ('b -> 'a -> 'b) -> 'a Beluga_syntax.Synint.LF.ctx -> 'b

General eliminator for contexts.

val map : ('a -> 'b) -> 'a Beluga_syntax.Synint.LF.ctx -> 'b Beluga_syntax.Synint.LF.ctx

Lift a function into contexts.

val to_list_map_rev : 'a Beluga_syntax.Synint.LF.ctx -> ('a Beluga_syntax.Synint.LF.ctx -> 'a -> 'b) -> 'b list

val to_list_map : 'a Beluga_syntax.Synint.LF.ctx -> ('a Beluga_syntax.Synint.LF.ctx -> 'a -> 'b) -> 'b list

val to_list_rev : 'a Beluga_syntax.Synint.LF.ctx -> 'a list

val to_list : 'a Beluga_syntax.Synint.LF.ctx -> 'a list

val to_sublist : 'a Beluga_syntax.Synint.LF.ctx -> ('a Beluga_syntax.Synint.LF.ctx * 'a) list

val to_sublist_rev : 'a Beluga_syntax.Synint.LF.ctx -> ('a Beluga_syntax.Synint.LF.ctx * 'a) list

val of_list_map : 'a list -> ('a -> 'b) -> 'b Beluga_syntax.Synint.LF.ctx

val of_list_map_rev : 'a list -> ('a -> 'b) -> 'b Beluga_syntax.Synint.LF.ctx

val of_list : 'a list -> 'a Beluga_syntax.Synint.LF.ctx

val of_list_rev : 'a list -> 'a Beluga_syntax.Synint.LF.ctx

val iter : 'a Beluga_syntax.Synint.LF.ctx -> ('a Beluga_syntax.Synint.LF.ctx -> 'a -> unit) -> unit

val iter' : 'a Beluga_syntax.Synint.LF.ctx -> ('a -> unit) -> unit

val iter_rev : 'a Beluga_syntax.Synint.LF.ctx -> ('a Beluga_syntax.Synint.LF.ctx -> 'a -> unit) -> unit

val iter_rev' : 'a Beluga_syntax.Synint.LF.ctx -> ('a -> unit) -> unit

val length : 'a Beluga_syntax.Synint.LF.ctx -> int

val dctxLength : Beluga_syntax.Synint.LF.dctx -> int

val find : 'a Beluga_syntax.Synint.LF.ctx -> ('a Beluga_syntax.Synint.LF.ctx -> 'a -> bool) -> 'a option

val find' : 'a Beluga_syntax.Synint.LF.ctx -> ('a -> bool) -> 'a option

val find_rev : 'a Beluga_syntax.Synint.LF.ctx -> ('a Beluga_syntax.Synint.LF.ctx -> 'a -> bool) -> 'a option

val find_rev' : 'a Beluga_syntax.Synint.LF.ctx -> ('a -> bool) -> 'a option

val find_index : 'a Beluga_syntax.Synint.LF.ctx -> ('a Beluga_syntax.Synint.LF.ctx -> 'a -> bool) -> int option

val find_index' : 'a Beluga_syntax.Synint.LF.ctx -> ('a -> bool) -> int option

val find_index_rev : 'a Beluga_syntax.Synint.LF.ctx -> ('a Beluga_syntax.Synint.LF.ctx -> 'a -> bool) -> int option

val find_index_rev' : 'a Beluga_syntax.Synint.LF.ctx -> ('a -> bool) -> int option

val find_with_index : 'a Beluga_syntax.Synint.LF.ctx -> ('a Beluga_syntax.Synint.LF.ctx -> ('a * int) -> bool) -> ('a * int) option

Finds the leftmost element of the context, together with its index, satisfying the given predicate

val find_with_index' : 'a Beluga_syntax.Synint.LF.ctx -> (('a * int) -> bool) -> ('a * int) option

val find_with_index_rev : 'a Beluga_syntax.Synint.LF.ctx -> ('a Beluga_syntax.Synint.LF.ctx -> ('a * int) -> bool) -> ('a * int) option

val find_with_index_rev' : 'a Beluga_syntax.Synint.LF.ctx -> (('a * int) -> bool) -> ('a * int) option

val getNameDCtx : Beluga_syntax.Synint.LF.dctx -> int -> Beluga_syntax.Name.t

val getNameMCtx : Beluga_syntax.Synint.LF.mctx -> int -> Beluga_syntax.Name.t

val getNameCtx : Beluga_syntax.Synint.Comp.gctx -> int -> Beluga_syntax.Name.t

val projectCtxIntoDctx : Beluga_syntax.Synint.LF.typ_decl Beluga_syntax.Synint.LF.ctx -> Beluga_syntax.Synint.LF.dctx

val splitContextVariable : Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.typ_decl -> Beluga_syntax.Synint.LF.dctx

val emptyContextVariable : Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.dctx

val lookup' : 'a Beluga_syntax.Synint.LF.ctx -> int -> 'a option

val lookup_inductivity : Beluga_syntax.Synint.LF.mctx -> int -> (Beluga_syntax.Synint.LF.ctyp * Beluga_syntax.Inductivity.t) option

Looks up an index in a meta-context, giving the type and `inductivity` field.

Warning: This function does not MShift the type it returns; typically you should do this so the type makes sense in the whole context Delta.

Warning: returns None if the index is out of bounds OR if the context declaration does not assign a type (i.e. it's a DeclOpt). If you need to distinguish these cases, then you should use lookup' and match on the ctyp_decl yourself.

val lookup : Beluga_syntax.Synint.Comp.gctx -> int -> Beluga_syntax.Synint.Comp.typ option

val lookupSchema : Beluga_syntax.Synint.LF.mctx -> int -> Beluga_syntax.Id.cid_schema

val lookupCtxVar : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.ctx_var -> Beluga_syntax.Name.t * Beluga_syntax.Id.cid_schema

val lookupCtxVarSchema : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.ctx_var -> Beluga_syntax.Id.cid_schema

val rename_mctx : Beluga_syntax.Name.t -> Beluga_syntax.Name.t -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.mctx option

Renames a variable in a meta-context. Returns None if there is no such variable, and otherwise returns the adjusted context with the name changed.

val rename_gctx : Beluga_syntax.Name.t -> Beluga_syntax.Name.t -> Beluga_syntax.Synint.Comp.gctx -> Beluga_syntax.Synint.Comp.gctx option

Renames a variable in a computational context. Returns None if there is no such variable, and otherwise returns the adjusted context with the name changed.

val concat : 'a Beluga_syntax.Synint.LF.ctx list -> 'a Beluga_syntax.Synint.LF.ctx

val names_of_dctx : Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Name.t list

Erases an LF context to a list of names. To be used in preparation for contextual name generation. See nameGen.ml.

val names_of_mctx : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Name.t list

Erases a meta-context to a list of names. To be used in preparation for contextual name generation. See nameGen.ml.

val names_of_gctx : Beluga_syntax.Synint.Comp.gctx -> Beluga_syntax.Name.t list

Erases a program context to a list of names. To be used in preparation for contextual name generation. See nameGen.ml.

val names_of_proof_state : Beluga_syntax.Synint.Comp.proof_state -> Beluga_syntax.Name.t list

Computes the names in a proof state. This is the names from the meta-context and the program context. To be used in preparation for contextual name generation. See nameGen.ml.

val is_null : Beluga_syntax.Synint.LF.dctx -> bool

val is_empty : 'a Beluga_syntax.Synint.LF.ctx -> bool

val steal_mctx_names : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.mctx

Given two convertible meta-contexts, rebuilds the first using the names in the second.

val steal_gctx_names : Beluga_syntax.Synint.Comp.gctx -> Beluga_syntax.Synint.Comp.gctx -> Beluga_syntax.Synint.Comp.gctx

Given two convertible computational contexts, rebuilds the first using the names in the second.