Module Beluga.ConvSigma

author Brigitte Pientka

type error =

| BlockInDctx of Beluga.Syntax.Int.LF.mctx * Beluga.Syntax.Int.LF.head * Beluga.Syntax.Int.LF.typ * Beluga.Syntax.Int.LF.dctx

type t

exception Error of Beluga_syntax.Common.Location.t * error

val fmt_ppr_conv_list : Stdlib.Format.formatter -> t -> unit

val strans_typ : Beluga.Syntax.Int.LF.mctx -> Beluga.Syntax.Int.LF.dctx -> Beluga.Syntax.Int.LF.tclo -> t -> Beluga.Syntax.Int.LF.typ

val strans_norm : Beluga.Syntax.Int.LF.mctx -> Beluga.Syntax.Int.LF.dctx -> Beluga.Syntax.Int.LF.nclo -> t -> Beluga.Syntax.Int.LF.normal

val strans_sub : Beluga.Syntax.Int.LF.mctx -> Beluga.Syntax.Int.LF.dctx -> Beluga.Syntax.Int.LF.sub -> t -> Beluga.Syntax.Int.LF.sub

val map : t -> Id.offset -> Id.offset

Translates a bound variable index according to the sigma conversion.

val flattenDCtx : Beluga.Syntax.Int.LF.mctx -> Beluga.Syntax.Int.LF.dctx -> Beluga.Syntax.Int.LF.dctx * t

val gen_proj_sub : t -> Beluga.Syntax.Int.LF.sub

val gen_tup_sub : t -> Beluga.Syntax.Int.LF.sub

val etaExpandMMVstr : Beluga_syntax.Common.Location.t -> Beluga.Syntax.Int.LF.mctx -> Beluga.Syntax.Int.LF.dctx -> Beluga.Syntax.Int.LF.tclo -> Beluga_syntax.Common.Plicity.t -> Name.t option -> Name.t list -> Beluga.Syntax.Int.LF.normal

Constructs a unification variable for the given tclo, strengthening its type.

val etaExpandMPVstr : Beluga_syntax.Common.Location.t -> Beluga.Syntax.Int.LF.mctx -> Beluga.Syntax.Int.LF.dctx -> Beluga.Syntax.Int.LF.tclo -> Beluga_syntax.Common.Plicity.t -> Name.t option -> Name.t list -> Beluga.Syntax.Int.LF.normal

Constructs a unification parameter variable for a given tclo, strengthening its type.

val gen_flattening : Beluga.Syntax.Int.LF.mctx -> Beluga.Syntax.Int.LF.dctx -> Beluga.Syntax.Int.LF.dctx * Beluga.Syntax.Int.LF.sub Stdlib.Lazy.t * Beluga.Syntax.Int.LF.sub Stdlib.Lazy.t

gen_flattening cD cPsi = (cPhi, lazy s_proj, lazy s_tup) Generates a flattened LF context cPhi in which all blocks present in cPsi have been decomposed. Packing and unpacking substitutions s_proj and s_tup are lazily computed to mediate between the contexts. Specifically: if cD |- cPsi ctx then cD |- cPhi ctx such that cD ; cPsi |- s_proj : cPhi cD ; cPhi |- s_tup : cPsi

The naming of s_proj and s_tup is based on what kind of terms these substitutions *contain*; e.g. s_proj contains projections, so it takes us from the block context to the flat context.