Module HoleId.Make

Construct a counter for hole id numbers. The reason this is wrapped in a dummy functor is that functor instantiation allows the get_next_key function to operate on a private integer reference, that can be made inaccessible outside of holes.ml

Parameters

Signature

val next : unit -> t

Generates the next hole ID.