Module Beluga.Holes

type lf_hole_info = {

cPsi : Beluga_syntax.Synint.LF.dctx;

lfGoal : Beluga_syntax.Synint.LF.tclo;

mutable lfSolution : Beluga_syntax.Synint.LF.nclo option;

}

type comp_hole_info = {

cG : Beluga_syntax.Synint.Comp.gctx;

compGoal : Beluga_syntax.Synint.Comp.typ * Beluga_syntax.Synint.LF.msub;

mutable compSolution : Beluga_syntax.Synint.Comp.exp option;

}

type _ hole_info =

| LFInfo : lf_hole_info hole_info

| CompInfo : comp_hole_info hole_info

type 'a hole = {

location : Beluga_syntax.Location.t;

name : Beluga_syntax.HoleId.name;

cD : Beluga_syntax.Synint.LF.mctx;
(*
"Context Delta", for metavariables.
*)

info : 'a;
(*
information specific to the hole type.
*)

}

A hole with a known kind of information.

type some_hole =

| Exists : 'a hole_info * 'a hole -> some_hole

Existential wrapper for 'a hole with a singleton type.

type lookup_strategy

A strategy for retrieving a hole.

val string_of_lookup_strategy : lookup_strategy -> string

Gets a string representation of the given strategy.

val parse_lookup_strategy : string -> lookup_strategy option

Parses a lookup strategy.

An integer gives the `by_id` strategy, whereas a non-integer gives the `by_name` strategy Returns None if the parsing fails.

val unsafe_parse_lookup_strategy : Beluga_syntax.Location.t -> string -> lookup_strategy

Parses a lookup strategy, throwing an exception on failure. The given location is used in the thrown exception.

val by_id : Beluga_syntax.HoleId.t -> lookup_strategy

Looks up a hole by its number. Use strategies with `get`.

val by_name : string -> lookup_strategy

Looks up a hole by its name. Use strategies with `get`.

type snapshot

val get_snapshot : unit -> snapshot

Gets a snapshot of the current hole ids.

val holes_since : snapshot -> (Beluga_syntax.HoleId.t * some_hole) list

Gets a list of holes added since the given snapshot. This can be used to whether an ad hoc expression contained any holes.

val catch : (unit -> 'a) -> (Beluga_syntax.HoleId.t * some_hole) list * 'a

Runs a function, and catches any new holes added as a result of running it.

val to_lf_hole : some_hole -> lf_hole_info hole option

Decides whether this is an LF hole.

val to_comp_hole : some_hole -> comp_hole_info hole option

Decides whether this is a computational hole.

val is_solved : some_hole -> bool

Decides whether this hole is solved.

val is_unsolved : some_hole -> bool

Decides whether this hole is unsolved.

val hole_is_named : some_hole -> bool

Decides whether a hole has a name (is not anonymous).

val none : unit -> bool

Checks whether the internal array of holes is empty.

val get : lookup_strategy -> (Beluga_syntax.HoleId.t * some_hole) option

Retrieves a single hole using the given strategy.

val unsafe_get : lookup_strategy -> Beluga_syntax.HoleId.t * some_hole

Retrieves a single hole using the given strategy, * raising NoSuchHole if the hole does not exist.

val lookup : string -> (Beluga_syntax.HoleId.t * some_hole) option

Looks up a hole, retrieving the hole itself and its number.

val loc_within : Beluga_syntax.Location.t -> Beluga_syntax.Location.t -> bool

Decides whether a location is contained within a location.

val count : unit -> int

Counts the number of holes.

val destroy_holes_within : Beluga_syntax.Location.t -> unit

Removes all holes contained in the given location.

val allocate : unit -> Beluga_syntax.HoleId.t

Allocates a new hole ID.

val assign : Beluga_syntax.HoleId.t -> some_hole -> unit

Assigns the hole to the given ID.

val clear : unit -> unit

Clears the hole arrays.

val list : unit -> (Beluga_syntax.HoleId.t * some_hole) list

Gets the current list of holes. The list is in sorted order of the hole locations in the source file.

val add_harpoon_subgoal : Beluga_syntax.Synint.Comp.open_subgoal -> unit

Adds a Harpoon subgoal to the internal list.

val get_harpoon_subgoals : unit -> Beluga_syntax.Synint.Comp.open_subgoal list

Gets the list of Harpoon subgoals.