Module Index.Indexer

include INDEXER with type state = Index_state.Indexing_state.state

include Support.Imperative_state.IMPERATIVE_STATE with type state = Index_state.Indexing_state.state

type state = Index_state.Indexing_state.state

val index_open_lf_kind : state -> Synext.lf_kind -> Beluga_syntax.Synapx.LF.kind

index_open_lf_kind state kind is kind with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are allowed in kind with respect to state.

val index_closed_lf_kind : state -> Synext.lf_kind -> Beluga_syntax.Synapx.LF.kind

index_closed_lf_kind state kind is kind with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are disallowed in kind with respect to state.

val index_open_lf_typ : state -> Synext.lf_typ -> Beluga_syntax.Synapx.LF.typ

index_open_lf_typ state typ is typ with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are allowed in typ with respect to state

val index_closed_lf_typ : state -> Synext.lf_typ -> Beluga_syntax.Synapx.LF.typ

index_closed_lf_typ state typ is typ with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are disallowed in typ with respect to state

val index_open_comp_kind : state -> Synext.comp_kind -> Beluga_syntax.Synapx.Comp.kind

index_open_comp_kind state kind is kind with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are allowed in kind with respect to state.

val index_closed_comp_kind : state -> Synext.comp_kind -> Beluga_syntax.Synapx.Comp.kind

index_closed_comp_kind state kind is kind with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are disallowed in kind with respect to state.

val index_open_comp_typ : state -> Synext.comp_typ -> Beluga_syntax.Synapx.Comp.typ

index_open_comp_typ state typ is typ with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are allowed in typ with respect to state.

val index_closed_comp_typ : state -> Synext.comp_typ -> Beluga_syntax.Synapx.Comp.typ

index_closed_comp_typ state typ is typ with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are disallowed in typ with respect to state.

val index_comp_expression : state -> Synext.comp_expression -> Beluga_syntax.Synapx.Comp.exp

index_comp_expression state expression is expression with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are disallowed in expression with respect to state.

val index_schema : state -> Synext.schema -> Beluga_syntax.Synapx.LF.schema

index_schema state schema is schema with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are allowed in schema with respect to state.

val index_comp_theorem : state -> Synext.comp_expression -> Beluga_syntax.Synapx.Comp.thm

index_comp_theorem state theorem is theorem with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are disallowed in theorem with respect to state.

val index_harpoon_proof : state -> Synext.harpoon_proof -> Beluga_syntax.Synapx.Comp.thm

index_harpoon_proof state proof is proof with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are disallowed in proof with respect to state.

val index_computation_typ_abbreviation : state -> Synext.comp_typ -> Synext.comp_kind -> Beluga_syntax.Synapx.Comp.typ * Beluga_syntax.Synapx.Comp.kind

index_computation_typ_abbreviation state typ kind is (typ', kind') where kind' is kind and typ' is typ, both with bound variables and constants replaced with de Bruijn indices and IDs respectively. Overall, free variables are disallowed with respect to state. Variables bound in Pi-kinds in kind count as bound in typ.

include Index_state.INDEXING_STATE with type state := state

include Support.Imperative_state.IMPERATIVE_STATE with type state := state

val traverse_tuple2 : state -> (state -> 'a1 -> 'b1) -> (state -> 'a2 -> 'b2) -> ('a1 * 'a2) -> 'b1 * 'b2

val traverse_tuple3 : state -> (state -> 'a1 -> 'b1) -> (state -> 'a2 -> 'b2) -> (state -> 'a3 -> 'b3) -> ('a1 * 'a2 * 'a3) -> 'b1 * 'b2 * 'b3

val traverse_list : state -> (state -> 'a -> 'b) -> 'a Support.List.t -> 'b Support.List.t

traverse_list state f [x1; x2; ...; xn] is [y1; y2; ...; yn] where yi = f state xi, and y1, y2, ..., yn are computed in order, meaning that y1 is computed first, then y2, etc.

val traverse_list1 : state -> (state -> 'a -> 'b) -> 'a Support.List1.t -> 'b Support.List1.t

val traverse_list2 : state -> (state -> 'a -> 'b) -> 'a Support.List2.t -> 'b Support.List2.t

val iter_list : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Support.List.t -> Stdlib.Unit.t

iter_list state f [x1; x2; ...; xn] is f state x1; f state x2; ...; f state xn.

val iter_list1 : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Support.List1.t -> Stdlib.Unit.t

val iter_list2 : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Support.List2.t -> Stdlib.Unit.t

val traverse_reverse_list : state -> (state -> 'a -> 'b) -> 'a Support.List.t -> 'b Support.List.t

traverse_reverse_list state f [x1; x2; ...; xn] is [y1; y2; ...; yn] where yi = f state xi, and y1, y2, ..., yn are computed in reverse order, meaning that yn is computed first, then y(n-1), etc.

val traverse_reverse_list1 : state -> (state -> 'a -> 'b) -> 'a Support.List1.t -> 'b Support.List1.t

val traverse_reverse_list2 : state -> (state -> 'a -> 'b) -> 'a Support.List2.t -> 'b Support.List2.t

val iter_rev_list : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Support.List.t -> Stdlib.Unit.t

iter_rev_list state f [x1; x2; ...; xn] is f state xn; f state x(n-1); ...; f state x1.

val iter_rev_list1 : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Support.List1.t -> Stdlib.Unit.t

val iter_rev_list2 : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Support.List2.t -> Stdlib.Unit.t

val traverse_option : state -> (state -> 'a -> 'b) -> 'a Support.Option.t -> 'b Support.Option.t

val iter_option : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Support.Option.t -> Stdlib.Unit.t

val seq_list : state -> (state -> 'a) Support.List.t -> 'a Support.List.t

val seq_list1 : state -> (state -> 'a) Support.List1.t -> 'a Support.List1.t

val iter_seq : state -> (state -> Stdlib.Unit.t) Support.List.t -> Stdlib.Unit.t

val fresh_identifier : state -> Beluga_syntax.Identifier.t

fresh_identifier state is an identifier that is not bound in state. This is used in the indexing of arrow types to Pi-types, and to generate parameter identifiers for lambda abstractions that omit the parameter identifier.

In order to avoid potential captures, identifier is not a syntactically valid identifier. That is, identifier printed as is cannot be parsed into an identifier.

For debugging purposes, identifier is of the form "\"i#" where '#' is a positive integer. Ideally, fresh name generation would not be needed, but that would require that every AST node in the concrete syntax is directly expressible in the approximate and internal syntaxes.

FIXME: The indexing phase should not have to generate fresh identifiers.

val fresh_identifier_opt : state -> Beluga_syntax.Identifier.t Support.Option.t -> Beluga_syntax.Identifier.t

fresh_identifier_opt state identifier_opt is fresh_identifier state if identifier_opt = Option.None, and identifier if identifier_opt = Option.Some identifier.

Index of Constants

val index_of_constant : state -> Beluga_syntax.Qualified_identifier.t -> [ `Lf_type_constant of Beluga_syntax.Id.cid_typ | `Lf_term_constant of Beluga_syntax.Id.cid_term | `Computation_inductive_type_constant of Beluga_syntax.Id.cid_comp_typ | `Computation_stratified_type_constant of Beluga_syntax.Id.cid_comp_typ | `Computation_coinductive_type_constant of Beluga_syntax.Id.cid_comp_cotyp | `Computation_abbreviation_type_constant of Beluga_syntax.Id.cid_comp_typdef | `Computation_term_constructor of Beluga_syntax.Id.cid_comp_const | `Computation_term_destructor of Beluga_syntax.Id.cid_comp_dest | `Schema_constant of Beluga_syntax.Id.cid_schema | `Program_constant of Beluga_syntax.Id.cid_prog | `Module of Beluga_syntax.Id.module_id ]

val index_of_lf_type_constant : state -> Beluga_syntax.Qualified_identifier.t -> Beluga_syntax.Id.cid_typ

index_of_lf_type_constant state identifier is the constant ID of identifier in state if it is bound to an LF type-level constant. If identifier is bound to any other entry, then an exception is raised.

val index_of_lf_term_constant : state -> Beluga_syntax.Qualified_identifier.t -> Beluga_syntax.Id.cid_term

index_of_lf_term_constant state identifier is like index_of_lf_type_constant state identifier, but for LF term-level constants.

val index_of_comp_type_constant : state -> Beluga_syntax.Qualified_identifier.t -> Beluga_syntax.Id.cid_comp_typ

index_of_comp_type_constant state identifier is a lookup for the ID corresponding to identifier in state for computation-level inductive or stratified type constant.

val index_of_inductive_comp_constant : state -> Beluga_syntax.Qualified_identifier.t -> Beluga_syntax.Id.cid_comp_typ

val index_of_stratified_comp_constant : state -> Beluga_syntax.Qualified_identifier.t -> Beluga_syntax.Id.cid_comp_typ

val index_of_coinductive_comp_constant : state -> Beluga_syntax.Qualified_identifier.t -> Beluga_syntax.Id.cid_comp_cotyp

val index_of_abbreviation_comp_constant : state -> Beluga_syntax.Qualified_identifier.t -> Beluga_syntax.Id.cid_comp_typdef

val index_of_schema_constant : state -> Beluga_syntax.Qualified_identifier.t -> Beluga_syntax.Id.cid_schema

val index_of_comp_constructor : state -> Beluga_syntax.Qualified_identifier.t -> Beluga_syntax.Id.cid_comp_const

val index_of_comp_destructor : state -> Beluga_syntax.Qualified_identifier.t -> Beluga_syntax.Id.cid_comp_dest

val index_of_comp_program : state -> Beluga_syntax.Qualified_identifier.t -> Beluga_syntax.Id.cid_prog

Index of Variables

val index_of_lf_variable : state -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.offset

index_of_lf_variable state identifier is the LF-bound de Bruijn index of identifier in state.

If identifier is unbound, then an exception is raised.

val index_of_lf_variable_opt : state -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.offset Support.Option.t

index_of_lf_variable_opt state identifier is the LF-bound de Bruijn index of identifier in state.

If identifier is unbound, then offset_opt = Option.None.

If state is a pattern state, then offset_opt is additionally Option.None if it is not an inner pattern-bound variable.

val index_of_meta_variable : state -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.offset

index_of_meta_variable state identifier is the meta-level de Bruijn index of identifier in state.

If identifier is unbound, bound to a different entry than a meta-variable, or bound to a meta-variable of unknown de Bruijn level, then an exception is raised.

val index_of_meta_variable_opt : state -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.offset Support.Option.t

index_of_meta_variable_opt state identifier is the meta-level de Bruijn index of identifier in state.

If identifier is unbound, then offset_opt = Option.None.

If identifier is a bound meta-variable whose de Bruijn level is unknown (i.e. its binder is a meta-variable in a pattern), then offset_opt = Option.None.

If state is a pattern state, then offset_opt = Option.None if it is not an inner pattern-bound variable.

val index_of_parameter_variable : state -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.offset

val index_of_parameter_variable_opt : state -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.offset Support.Option.t

val index_of_substitution_variable : state -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.offset

val index_of_substitution_variable_opt : state -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.offset Support.Option.t

val index_of_context_variable : state -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.offset

val index_of_context_variable_opt : state -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.offset Support.Option.t

val index_of_comp_variable : state -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.offset

Binding Variables

val with_bound_lf_variable : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> (state -> 'a) -> 'a

with_bound_lf_variable state identifier m runs m in a state where identifier is a bound LF variable.

If state is a pattern state, then identifier is additionally considered as an inner bound variable.

val with_bound_meta_variable : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> (state -> 'a) -> 'a

val with_bound_parameter_variable : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> (state -> 'a) -> 'a

val with_bound_substitution_variable : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> (state -> 'a) -> 'a

val with_bound_context_variable : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> (state -> 'a) -> 'a

val with_bound_contextual_variable : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> (state -> 'a) -> 'a

with_bound_contextual_variable state identifier m runs m in a state where identifier is either a bound meta, parameter, substitution or context variable. This is necessary for mlam-expressions.

If state is a pattern state, then identifier is additionally considered as an inner bound variable.

val with_bound_comp_variable : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> (state -> 'a) -> 'a

val with_shifted_lf_context : state -> (state -> 'a) -> 'a

with_shifted_lf_context state m is like with_bound_lf_variable state _ m without adding any identifier in the namespace. That is, de Bruijn indices looked up in m with respect to the context of LF-bound variables are + 1 of what they were in state. This is used for omitted parameters to lambda terms, like \_. x.

val with_shifted_meta_context : state -> (state -> 'a) -> 'a

with_shifted_meta_context state m is like with_shifted_lf_context state m but for meta-level variables.

val with_shifted_comp_context : state -> (state -> 'a) -> 'a

with_shifted_comp_context state m is like with_shifted_lf_context state m but for computation-level variables.

val add_lf_type_constant : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.cid_typ -> Stdlib.Unit.t

val add_lf_term_constant : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.cid_term -> Stdlib.Unit.t

val add_schema_constant : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.cid_schema -> Stdlib.Unit.t

val add_inductive_computation_type_constant : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.cid_comp_typ -> Stdlib.Unit.t

val add_stratified_computation_type_constant : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.cid_comp_typ -> Stdlib.Unit.t

val add_coinductive_computation_type_constant : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.cid_comp_cotyp -> Stdlib.Unit.t

val add_abbreviation_computation_type_constant : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.cid_comp_typdef -> Stdlib.Unit.t

val add_computation_term_constructor : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.cid_comp_const -> Stdlib.Unit.t

val add_computation_term_destructor : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.cid_comp_dest -> Stdlib.Unit.t

val add_program_constant : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.cid_prog -> Stdlib.Unit.t

val add_module : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> Beluga_syntax.Id.module_id -> (state -> 'a) -> 'a

add_module state ?location module_identifier cid m starts a module, computes m, then stops the module. This effectively adds a new module with identifier module_identifier to state, along with the declarations added in m as declarations in that module.

val open_module : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Qualified_identifier.t -> Stdlib.Unit.t

open_module state ?location module_identifier opens the module module_identifier it is a bound module. This effectively adds all the declarations in that module to the current frame, but not as declarations (i.e., this opens the module like in OCaml).

val add_module_abbreviation : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Qualified_identifier.t -> Beluga_syntax.Identifier.t -> Stdlib.Unit.t

val with_frame : state -> (state -> 'a) -> 'a

with_frame state m runs m in a nested bindings frame that is discarded afterwards.

val with_parent_frame : state -> (state -> 'a) -> 'a

with_parent_frame state m runs m while ignoring the topmost frame.

This is used for indexing Harpoon proof scripts because Harpoon hypotheticals are already serialized with all the identifiers in the meta-level and computation-level contexts of the proof.

val with_bindings_checkpoint : state -> (state -> 'a) -> 'a

with_bindings_checkpoint state m runs m with a checkpoint on the bindings' state to backtrack to in case m raises an exception.

This is used in REPLs to safely run functions and recover the state in case of a raised exception during indexing.

This checkpoint mechanism does not perform a copy or snapshot of the referencing environment. You need to make sure that m does not pop too many scopes.

Pattern Variables

val with_free_variables_as_pattern_variables : state -> pattern:(state -> 'a) -> expression:(state -> 'a -> 'b) -> 'b

with_free_variables_as_pattern_variables state ~pattern ~expression runs pattern while keeping track of free and pattern variables, then runs expression with the free and pattern variables as bound variables.

val add_computation_pattern_variable : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> Stdlib.Unit.t

Free Variables

val add_free_lf_variable : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> Stdlib.Unit.t

add_free_lf_variable state identifier adds identifier as a free LF variable.

If identifier is a free variable in state of a different kind than LF variables, then an exception is raised.

val add_free_meta_variable : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> Stdlib.Unit.t

val add_free_parameter_variable : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> Stdlib.Unit.t

val add_free_substitution_variable : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> Stdlib.Unit.t

val add_free_context_variable : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> Stdlib.Unit.t

val with_bound_pattern_meta_variable : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> (state -> 'a) -> 'a

val with_bound_pattern_parameter_variable : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> (state -> 'a) -> 'a

val with_bound_pattern_substitution_variable : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> (state -> 'a) -> 'a

val with_bound_pattern_context_variable : state -> ?location:Beluga_syntax.Location.t -> Beluga_syntax.Identifier.t -> (state -> 'a) -> 'a

val allow_free_variables : state -> (state -> 'a) -> 'a

allow_free_variables state m runs m and discards the tracked free variables therein.

val disallow_free_variables : state -> (state -> 'a) -> 'a

disallow_free_variables state m runs m and raises an exception if m adds a free variable to the state.

Interoperability

val add_all_mctx : state -> Beluga_syntax.Synint.LF.mctx -> Stdlib.Unit.t

add_all_mctx state cD adds all contextual variables in cD.

val add_all_gctx : state -> Beluga_syntax.Synint.Comp.gctx -> Stdlib.Unit.t

add_all_gctx state cG adds all computation-level variables in cG.