Module type Index.INDEXER

include Support.Imperative_state.IMPERATIVE_STATE

type state

val traverse_tuple2 : state -> (state -> 'a1 -> 'b1) -> (state -> 'a2 -> 'b2) -> ('a1 * 'a2) -> 'b1 * 'b2

val traverse_tuple3 : state -> (state -> 'a1 -> 'b1) -> (state -> 'a2 -> 'b2) -> (state -> 'a3 -> 'b3) -> ('a1 * 'a2 * 'a3) -> 'b1 * 'b2 * 'b3

val traverse_list : state -> (state -> 'a -> 'b) -> 'a Support.List.t -> 'b Support.List.t

traverse_list state f [x1; x2; ...; xn] is [y1; y2; ...; yn] where yi = f state xi, and y1, y2, ..., yn are computed in order, meaning that y1 is computed first, then y2, etc.

val traverse_list1 : state -> (state -> 'a -> 'b) -> 'a Support.List1.t -> 'b Support.List1.t

val traverse_list2 : state -> (state -> 'a -> 'b) -> 'a Support.List2.t -> 'b Support.List2.t

val iter_list : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Support.List.t -> Stdlib.Unit.t

iter_list state f [x1; x2; ...; xn] is f state x1; f state x2; ...; f state xn.

val iter_list1 : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Support.List1.t -> Stdlib.Unit.t

val iter_list2 : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Support.List2.t -> Stdlib.Unit.t

val traverse_reverse_list : state -> (state -> 'a -> 'b) -> 'a Support.List.t -> 'b Support.List.t

traverse_reverse_list state f [x1; x2; ...; xn] is [y1; y2; ...; yn] where yi = f state xi, and y1, y2, ..., yn are computed in reverse order, meaning that yn is computed first, then y(n-1), etc.

val traverse_reverse_list1 : state -> (state -> 'a -> 'b) -> 'a Support.List1.t -> 'b Support.List1.t

val traverse_reverse_list2 : state -> (state -> 'a -> 'b) -> 'a Support.List2.t -> 'b Support.List2.t

val iter_rev_list : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Support.List.t -> Stdlib.Unit.t

iter_rev_list state f [x1; x2; ...; xn] is f state xn; f state x(n-1); ...; f state x1.

val iter_rev_list1 : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Support.List1.t -> Stdlib.Unit.t

val iter_rev_list2 : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Support.List2.t -> Stdlib.Unit.t

val traverse_option : state -> (state -> 'a -> 'b) -> 'a Support.Option.t -> 'b Support.Option.t

val iter_option : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Support.Option.t -> Stdlib.Unit.t

val seq_list : state -> (state -> 'a) Support.List.t -> 'a Support.List.t

val seq_list1 : state -> (state -> 'a) Support.List1.t -> 'a Support.List1.t

val iter_seq : state -> (state -> Stdlib.Unit.t) Support.List.t -> Stdlib.Unit.t

val index_open_lf_kind : state -> Synext.lf_kind -> Beluga_syntax.Synapx.LF.kind

index_open_lf_kind state kind is kind with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are allowed in kind with respect to state.

val index_closed_lf_kind : state -> Synext.lf_kind -> Beluga_syntax.Synapx.LF.kind

index_closed_lf_kind state kind is kind with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are disallowed in kind with respect to state.

val index_open_lf_typ : state -> Synext.lf_typ -> Beluga_syntax.Synapx.LF.typ

index_open_lf_typ state typ is typ with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are allowed in typ with respect to state

val index_closed_lf_typ : state -> Synext.lf_typ -> Beluga_syntax.Synapx.LF.typ

index_closed_lf_typ state typ is typ with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are disallowed in typ with respect to state

val index_open_comp_kind : state -> Synext.comp_kind -> Beluga_syntax.Synapx.Comp.kind

index_open_comp_kind state kind is kind with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are allowed in kind with respect to state.

val index_closed_comp_kind : state -> Synext.comp_kind -> Beluga_syntax.Synapx.Comp.kind

index_closed_comp_kind state kind is kind with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are disallowed in kind with respect to state.

val index_open_comp_typ : state -> Synext.comp_typ -> Beluga_syntax.Synapx.Comp.typ

index_open_comp_typ state typ is typ with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are allowed in typ with respect to state.

val index_closed_comp_typ : state -> Synext.comp_typ -> Beluga_syntax.Synapx.Comp.typ

index_closed_comp_typ state typ is typ with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are disallowed in typ with respect to state.

val index_comp_expression : state -> Synext.comp_expression -> Beluga_syntax.Synapx.Comp.exp

index_comp_expression state expression is expression with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are disallowed in expression with respect to state.

val index_schema : state -> Synext.schema -> Beluga_syntax.Synapx.LF.schema

index_schema state schema is schema with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are allowed in schema with respect to state.

val index_comp_theorem : state -> Synext.comp_expression -> Beluga_syntax.Synapx.Comp.thm

index_comp_theorem state theorem is theorem with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are disallowed in theorem with respect to state.

val index_harpoon_proof : state -> Synext.harpoon_proof -> Beluga_syntax.Synapx.Comp.thm

index_harpoon_proof state proof is proof with bound variables and constants replaced with de Bruijn indices and IDs respectively. Free variables are disallowed in proof with respect to state.

val index_computation_typ_abbreviation : state -> Synext.comp_typ -> Synext.comp_kind -> Beluga_syntax.Synapx.Comp.typ * Beluga_syntax.Synapx.Comp.kind

index_computation_typ_abbreviation state typ kind is (typ', kind') where kind' is kind and typ' is typ, both with bound variables and constants replaced with de Bruijn indices and IDs respectively. Overall, free variables are disallowed with respect to state. Variables bound in Pi-kinds in kind count as bound in typ.