Module Beluga.Lfcheck

module P = Prettyint.DefaultPrinter

module R = Store.Cid.DefaultRenderer

val dprintf : Support.Debug.Fmt.fmt Support.Debug.io Support.Debug.io

val dprint : (unit -> string) Support.Debug.io

val dprnt : string Support.Debug.io

module Unify = Unify.EmptyTrail

module S = Substitution

type error =

| CtxVarMisCheck of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.tclo * Beluga_syntax.Name.t * Beluga_syntax.Synint.LF.schema

| CtxVarMismatch of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.LF.ctx_var * Beluga_syntax.Name.t * Beluga_syntax.Synint.LF.schema

| CtxVarDiffer of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.LF.ctx_var * Beluga_syntax.Synint.LF.ctx_var

| CheckError of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.nclo * Beluga_syntax.Synint.LF.tclo

| TupleArity of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.nclo * Beluga_syntax.Synint.LF.trec_clo

| SigmaMismatch of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.trec_clo * Beluga_syntax.Synint.LF.trec_clo

| KindMismatch of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.sclo * Beluga_syntax.Synint.LF.kind * Beluga_syntax.Synint.LF.sub

| TypMismatch of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.nclo * Beluga_syntax.Synint.LF.tclo * Beluga_syntax.Synint.LF.tclo

| IllTypedSub of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.sub * Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.normal option

| SpineIllTyped of int * int

| LeftoverFV of Beluga_syntax.Name.t

| ParamVarInst of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.tclo

| CtxHatMismatch of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.dctx_hat * Beluga_syntax.Location.t * Beluga_syntax.Synint.LF.mfront

| IllTypedMetaObj of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.LF.clobj * Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.cltyp

| TermWhenVar of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.normal

| SubWhenRen of Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.sub

| MissingType of string

exception Error of Beluga_syntax.Location.t * error

val throw : Beluga_syntax.Location.t -> error -> 'a

val error_printer : error -> Support.Format.formatter -> unit

exception SpineMismatch

val convPrefixCtx : Beluga_syntax.Synint.LF.typ_decl Beluga_syntax.Synint.LF.ctx -> Beluga_syntax.Synint.LF.typ_decl Beluga_syntax.Synint.LF.ctx -> bool

val dot_k : Beluga_syntax.Synint.LF.sub -> int -> Beluga_syntax.Synint.LF.sub

val convPrefixTypRec : (Beluga_syntax.Synint.LF.typ_rec * Beluga_syntax.Synint.LF.sub) -> (Beluga_syntax.Synint.LF.typ_rec * Beluga_syntax.Synint.LF.sub) -> bool

val prefixSchElem : Beluga_syntax.Synint.LF.sch_elem -> Beluga_syntax.Synint.LF.sch_elem -> bool

val ctxToSub' : Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.sub

val checkW : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.nclo -> Beluga_syntax.Synint.LF.tclo -> unit

val check : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.nclo -> Beluga_syntax.Synint.LF.tclo -> unit

val checkTuple : Syncom.Location.t -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> (Beluga_syntax.Synint.LF.tuple * Beluga_syntax.Synint.LF.sub) -> (Beluga_syntax.Synint.LF.typ_rec * Beluga_syntax.Synint.LF.sub) -> unit

val syn : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.nclo -> Beluga_syntax.Synint.LF.tclo

val inferHead : Syncom.Location.t -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.head -> Beluga_syntax.Synint.LF.svar_class -> Beluga_syntax.Synint.LF.typ

val canAppear : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.head -> (Beluga_syntax.Synint.LF.typ * Beluga_syntax.Synint.LF.sub) -> Beluga_syntax.Location.t -> bool

val checkSub : Syncom.Location.t -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> ?tM_opt:Beluga_syntax.Synint.LF.normal option -> Beluga_syntax.Synint.LF.sub -> Beluga_syntax.Synint.LF.svar_class -> Beluga_syntax.Synint.LF.dctx -> unit

checkSub loc cD cPsi s cPsi' = ()

Succeeds iff cD ; cPsi |- s : cPsi'

The optional term tM_opt, which is the term to which the substitution is being applied, is added to error messages to make them more informative. If tM_opt = Some tM then require tM be well-formed in cD; cPsi'

val synKSpine : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> (Beluga_syntax.Synint.LF.spine * Beluga_syntax.Synint.LF.sub) -> (Beluga_syntax.Synint.LF.kind * Beluga_syntax.Synint.LF.sub) -> Beluga_syntax.Synint.LF.kind * Beluga_syntax.Synint.LF.sub

val checkTyp' : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.tclo -> unit

val checkTyp : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.tclo -> unit

val checkTypRec : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> (Beluga_syntax.Synint.LF.typ_rec * Beluga_syntax.Synint.LF.sub) -> unit

val checkKind : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.kind -> unit

val checkDec : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> (Beluga_syntax.Synint.LF.typ_decl * Beluga_syntax.Synint.LF.sub) -> unit

val checkDCtx : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> unit

val projectCtxIntoDctx : Beluga_syntax.Synint.LF.typ_decl Beluga_syntax.Synint.LF.ctx -> Beluga_syntax.Synint.LF.dctx

val checkTypeAgainstSchema : Beluga_syntax.Location.t -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.typ -> Beluga_syntax.Name.t -> Beluga_syntax.Synint.LF.sch_elem list -> Beluga_syntax.Synint.LF.typ_rec * Beluga_syntax.Synint.LF.sub

val instanceOfSchElem : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> (Beluga_syntax.Synint.LF.typ * Beluga_syntax.Synint.LF.sub) -> Beluga_syntax.Synint.LF.sch_elem -> Beluga_syntax.Synint.LF.typ_rec * Beluga_syntax.Synint.LF.sub

val checkTypeAgainstSchemaProj : Beluga_syntax.Location.t -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.head -> Beluga_syntax.Synint.LF.typ -> Beluga_syntax.Name.t -> Beluga_syntax.Synint.LF.sch_elem list -> Beluga_syntax.Synint.LF.typ_rec * Beluga_syntax.Synint.LF.sub

val existsInstOfSchElemProj : Beluga_syntax.Location.t -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.tclo -> (Beluga_syntax.Synint.LF.head * int * int) -> Beluga_syntax.Synint.LF.sch_elem -> Beluga_syntax.Synint.LF.typ_rec * Beluga_syntax.Synint.LF.sub

val instanceOfSchElemProj : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.tclo -> (Beluga_syntax.Synint.LF.head * int) -> Beluga_syntax.Synint.LF.sch_elem -> Beluga_syntax.Synint.LF.typ_rec * Beluga_syntax.Synint.LF.sub

val checkElementAgainstElement : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.sch_elem -> Beluga_syntax.Synint.LF.sch_elem -> bool

val checkElementAgainstSchema : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.sch_elem -> Beluga_syntax.Synint.LF.sch_elem list -> bool

val checkSchema : Beluga_syntax.Location.t -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Name.t -> Beluga_syntax.Synint.LF.schema -> unit

val subsumes : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.ctx_var -> Beluga_syntax.Synint.LF.ctx_var -> bool

val elemPostfix : (Beluga_syntax.Synint.LF.typ_rec * Beluga_syntax.Synint.LF.sub) -> (Beluga_syntax.Synint.LF.typ_rec * Beluga_syntax.Synint.LF.sub) -> (Beluga_syntax.Synint.LF.typ_rec * Beluga_syntax.Synint.LF.sub) option

val checkSchemaWf : Beluga_syntax.Synint.LF.schema -> unit

val checkClObj : Beluga_syntax.Synint.LF.mctx -> Syncom.Location.t -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.clobj -> (Beluga_syntax.Synint.LF.cltyp * Beluga_syntax.Synint.LF.msub) -> unit

val checkMetaObj : Beluga_syntax.Synint.LF.mctx -> (Beluga_syntax.Location.t * Beluga_syntax.Synint.LF.mfront) -> (Beluga_syntax.Synint.LF.ctyp * Beluga_syntax.Synint.LF.msub) -> unit

val checkMSub : Syncom.Location.t -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.msub -> Beluga_syntax.Synint.LF.mctx -> unit