Module Beluga.Logic

type goal

type comp_goal

type query

type mquery

type comp_res

module Options : sig ... end

module Convert : sig ... end

module Index : sig ... end

module Frontend : sig ... end

type bound = int option

module Solver : sig ... end

module CSolver : sig ... end

val storeQuery : Beluga_syntax.Name.t option -> (Beluga_syntax.Synint.LF.typ * Beluga_syntax.Id.offset) -> Beluga_syntax.Synint.LF.mctx -> bound -> bound -> unit

val runLogic : unit -> unit

val runLogicOn : Beluga_syntax.Name.t option -> (Beluga_syntax.Synint.LF.typ * Beluga_syntax.Id.offset) -> Beluga_syntax.Synint.LF.mctx -> bound -> bound -> unit

Clears the local storage of the logic programming engine, * and loads the LF signature.

val prepare : unit -> unit

Clears the local storage of the logic programming engine, * and loads the LF signature.