Module Beluga.Name

type t

A `name' represents the original named representation of a variable. These are primarily used in parsing and pretty- printing. Names should never be constructed directly. See `mk_name'.

type max_usage = [

| `used of int option

| `unused

]

val base_name : t -> string

val get_module : t -> string list

val mk_blank : Beluga_syntax.Location.t option -> t

Generate a blank name (underscore).

val inc_hint_cnt : int option -> int option

val gen_fresh_name : t list -> t -> t

val location : t -> Beluga_syntax.Location.t

Retrieves the location of this identifier. For identifiers generated internally, this will be a ghost.

val max_usage : t list -> string -> max_usage

Finds the maximum number used for the given name hint in the given context. Returns None if the name hint is unused.

val modify_number : (int option -> int option) -> t -> t

Change the number of a variable.

type name_guide =

| NoName

| MVarName of (unit -> string) option

| SVarName of (unit -> string) option

| PVarName of (unit -> string) option

| BVarName of (unit -> string) option

| SomeName of t

| SomeString of string

val mk_name : ?location:Beluga_syntax.Location.t -> ?modules:string list -> name_guide -> t

Smart constructor for t.

mk_name generates a t with a guaranteed unique string.

val make_from_identifier : Beluga_syntax.Identifier.t -> t

val string_of_name : t -> string

val pp_list : Support.Format.formatter -> t list -> unit

Prints a list of space-separated names. This printer does not open a box!

Instances

Decide whether two names represent the same thing. This simply compares the string representations of the names. This corresponds with a user's intuition about when names are equal, and ignores all name generation hinting built in to the t type.

include Support.Eq.EQ with type t := t

val equal : t -> t -> bool

equal a b is true if and only if a and b are equal. This should satisfy the following properties:

Reflexivity: equal a a = true

Symmetry: equal a b is equivalent to equal b a

Transitivity: if equal a b = true and equal b c = true, then equal a c = true

val (=) : t -> t -> bool

Operator alias of equal.

val (<>) : t -> t -> bool

Negation of operator (=).

include Support.Show.SHOW with type t := t

val pp : Support.Format.formatter -> t -> unit

pp ppf t emits t pretty-printed to ppf.

val show : t -> string

show t pretty-prints t to a string.

include Support.Hash.HASH with type t := t

val hash : t -> int

hash e is the hash code computed for e.

If e1 and e2 as equal, then hash e1 and hash e2 must be equal.