Module Options.Monitor

val on : bool Stdlib.ref

on: is true if we call the interpreter with argument: +t The result is to print the timings to stdout when signature reconstruction ends.

val onf : bool Stdlib.ref

onf: is true if we call the interpreter with argument: +tfile The result is to dump the timing to a file when signature reconstruction ends.

val is_active : unit -> bool

Decides whether either timing mode (stdout or file) is active