Module Beluga.Order

module I = Beluga_syntax.Synint.Comp

type mutual =

| Empty

| LE of Beluga_syntax.Id.cid_prog * mutual

| LT of Beluga_syntax.Id.cid_prog * mutual

type dec =

| Dec of I.order * mutual

val list_of_order : I.order -> int list option

Converts the order to a list of argument positions If the order is too complicated, returns None.