Module Beluga.Prettyint

val dprnt : string Support.Debug.io

module PC = Printer.Control

val ident_regexp : Str.regexp

val is_name_syntactically_valid : Beluga_syntax.Name.t -> bool

is_name_syntactically_valid name is a heuristic for determining whether name is syntactically valid. If that is not the case, then the name must not have been user-defined, so it was generated during signature reconstruction.

module Make (R : Store.Cid.RENDERER) : Printer.Int.T

module DefaultPrinter : sig ... end