Module Beluga.Reconstruct

val solve_fvarCnstr : Lfrecon.reconType -> unit

val reset_fvarCnstr : unit -> unit

type case_type

Type of case to reconstruct. Either a plain data object, or an index object. In the case of an index object, we store the scrutinee in order to correctly implement dependent pattern matching.

val case_type : Beluga_syntax.Int.Comp.pattern Stdlib.Lazy.t -> Beluga_syntax.Int.Comp.exp -> case_type

Analyzes the given case scrutinee to decide what kind of case we have. The case type is used by other elaboration functions to generate appropriate substitutions. (Specifically, see synPatRefine, which is the main consumer of case_type.)

The given pattern is passed lazily as it is only needed when the synthesizable expression happens to be a box. (This is related to the implementation of dependent pattern matching.) If you know that the synthesizable expression is definitely not a box, then you can pass lazy (Error.raise_violation "appropriate message") as the pattern.

val map_case_type : ((Beluga_syntax.Int.Comp.pattern * Beluga_syntax.Int.Comp.meta_obj) -> Beluga_syntax.Int.Comp.pattern * Beluga_syntax.Int.Comp.meta_obj) -> case_type -> case_type

Transforms a case_type if it was actually an index object.

val synPatRefine : Beluga_syntax.Location.t -> case_type -> (Beluga_syntax.Int.LF.mctx * Beluga_syntax.Int.LF.mctx) -> Beluga_syntax.Int.LF.msub -> (Beluga_syntax.Int.Comp.typ * Beluga_syntax.Int.Comp.typ) -> Beluga_syntax.Int.LF.msub * Beluga_syntax.Int.LF.msub * Beluga_syntax.Int.LF.mctx

Constructs the refinement substitution for a case analysis. If * cD; cG |- tau_s <= type * cD'; cG' |- tau_p <= type * cD' |- t : cD * if caseT = IndexObj C then

tau_s = U

cD; cG |- C <= U then synPatRefine' loc caseT (cD, cD') t (tau_s, tau_p) = t', t1', cD'' such that * cD'' |- t' : cD * cD'' |- t1' : cD'

Remark: * cD'' is the meta-context in which the branch of the case should be elaborated. * t' is just t1' composed with t.

val kind : Beluga_syntax.Apx.LF.kind -> Beluga_syntax.Int.LF.kind

val typ : Lfrecon.reconType -> Beluga_syntax.Apx.LF.typ -> Beluga_syntax.Int.LF.typ

val schema : Beluga_syntax.Apx.LF.schema -> Beluga_syntax.Int.LF.schema

val mctx : Beluga_syntax.Apx.LF.mctx -> Beluga_syntax.Int.LF.mctx

val compkind : Beluga_syntax.Apx.Comp.kind -> Beluga_syntax.Int.Comp.kind

val comptyp : Beluga_syntax.Apx.Comp.typ -> Beluga_syntax.Int.Comp.typ

val comptyp_cD : Beluga_syntax.Int.LF.mctx -> Beluga_syntax.Apx.Comp.typ -> Beluga_syntax.Int.Comp.typ

val comptypdef : Beluga_syntax.Location.t -> Beluga_syntax.Name.t -> (Beluga_syntax.Apx.Comp.typ * Beluga_syntax.Apx.Comp.kind) -> (Beluga_syntax.Int.LF.mctx * Beluga_syntax.Int.Comp.typ) * Beluga_syntax.Id.offset * Beluga_syntax.Int.Comp.kind

val elExp : Beluga_syntax.Int.LF.mctx -> Beluga_syntax.Int.Comp.gctx -> Beluga_syntax.Apx.Comp.exp -> (Beluga_syntax.Int.Comp.typ * Beluga_syntax.Int.LF.msub) -> Beluga_syntax.Int.Comp.exp

val elExp' : Beluga_syntax.Int.LF.mctx -> Beluga_syntax.Int.Comp.gctx -> Beluga_syntax.Apx.Comp.exp -> Beluga_syntax.Int.Comp.exp * Beluga_syntax.Int.Comp.tclo

val exp : Beluga_syntax.Int.Comp.gctx -> Beluga_syntax.Apx.Comp.exp -> (Beluga_syntax.Int.Comp.typ * Beluga_syntax.Int.LF.msub) -> Beluga_syntax.Int.Comp.exp

val exp' : Beluga_syntax.Int.Comp.gctx -> Beluga_syntax.Apx.Comp.exp -> Beluga_syntax.Int.Comp.exp * Beluga_syntax.Int.Comp.tclo

val thm : Beluga_syntax.Int.Comp.gctx -> Beluga_syntax.Apx.Comp.thm -> (Beluga_syntax.Int.Comp.typ * Beluga_syntax.Int.LF.msub) -> Beluga_syntax.Int.Comp.thm

val numeric_order : Beluga_syntax.Int.Comp.typ -> Beluga_syntax.Int.Comp.order -> Beluga_syntax.Int.Comp.order