Module Beluga.Recsgn

type leftover_vars = (Abstract.free_var Beluga.Syntax.Int.LF.ctx * Beluga_syntax.Common.Location.t) list

val apply_global_pragmas : Beluga.Syntax.Ext.Sgn.decl list -> Beluga.Syntax.Ext.Sgn.decl list

Processes global pragmas, which must appear at the beginning of the list, and returns the remaining declarations.

val recSgnDecls : Beluga.Syntax.Ext.Sgn.decl list -> Beluga.Syntax.Int.Sgn.decl list * leftover_vars option

val fmt_ppr_leftover_vars : Stdlib.Format.formatter -> leftover_vars -> unit

val ppr_leftover_vars : leftover_vars -> unit