Module Recsgn_state.Make_signature_reconstruction_state

Parameters

module Index_state : Index_state.INDEXING_STATE

module Index : Index.INDEXER with type state = Index_state.state

Signature

include SIGNATURE_RECONSTRUCTION_STATE

include Support.Imperative_state.IMPERATIVE_STATE

type state

val traverse_tuple2 : state -> (state -> 'a1 -> 'b1) -> (state -> 'a2 -> 'b2) -> ('a1 * 'a2) -> 'b1 * 'b2

val traverse_tuple3 : state -> (state -> 'a1 -> 'b1) -> (state -> 'a2 -> 'b2) -> (state -> 'a3 -> 'b3) -> ('a1 * 'a2 * 'a3) -> 'b1 * 'b2 * 'b3

val traverse_list : state -> (state -> 'a -> 'b) -> 'a Support.List.t -> 'b Support.List.t

traverse_list state f [x1; x2; ...; xn] is [y1; y2; ...; yn] where yi = f state xi, and y1, y2, ..., yn are computed in order, meaning that y1 is computed first, then y2, etc.

val traverse_list1 : state -> (state -> 'a -> 'b) -> 'a Support.List1.t -> 'b Support.List1.t

val traverse_list2 : state -> (state -> 'a -> 'b) -> 'a Support.List2.t -> 'b Support.List2.t

val iter_list : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Support.List.t -> Stdlib.Unit.t

iter_list state f [x1; x2; ...; xn] is f state x1; f state x2; ...; f state xn.

val iter_list1 : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Support.List1.t -> Stdlib.Unit.t

val iter_list2 : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Support.List2.t -> Stdlib.Unit.t

val traverse_reverse_list : state -> (state -> 'a -> 'b) -> 'a Support.List.t -> 'b Support.List.t

traverse_reverse_list state f [x1; x2; ...; xn] is [y1; y2; ...; yn] where yi = f state xi, and y1, y2, ..., yn are computed in reverse order, meaning that yn is computed first, then y(n-1), etc.

val traverse_reverse_list1 : state -> (state -> 'a -> 'b) -> 'a Support.List1.t -> 'b Support.List1.t

val traverse_reverse_list2 : state -> (state -> 'a -> 'b) -> 'a Support.List2.t -> 'b Support.List2.t

val iter_rev_list : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Support.List.t -> Stdlib.Unit.t

iter_rev_list state f [x1; x2; ...; xn] is f state xn; f state x(n-1); ...; f state x1.

val iter_rev_list1 : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Support.List1.t -> Stdlib.Unit.t

val iter_rev_list2 : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Support.List2.t -> Stdlib.Unit.t

val traverse_option : state -> (state -> 'a -> 'b) -> 'a Support.Option.t -> 'b Support.Option.t

val iter_option : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Support.Option.t -> Stdlib.Unit.t

val seq_list : state -> (state -> 'a) Support.List.t -> 'a Support.List.t

val seq_list1 : state -> (state -> 'a) Support.List1.t -> 'a Support.List1.t

val iter_seq : state -> (state -> Stdlib.Unit.t) Support.List.t -> Stdlib.Unit.t

val get_leftover_vars : state -> (Abstract.free_var Synint.LF.ctx * Beluga_syntax.Syncom.Location.t) Support.List.t

val add_leftover_vars : state -> Abstract.free_var Synint.LF.ctx -> Beluga_syntax.Syncom.Location.t -> Stdlib.Unit.t

val with_disabled_lf_strengthening : state -> location:Beluga_syntax.Syncom.Location.t -> (state -> 'a) -> 'a

disable_lf_strengthening state ~location disables the strengthening of LF terms and types during LF reconstruction.

location is the location to use for error-reporting in case of failure to perform disable_lf_strengthening ~location.

val with_warn_on_coverage_error : state -> location:Beluga_syntax.Syncom.Location.t -> (state -> 'a) -> 'a

with_warn_on_coverage_error state ~location sets the error-reporting mode of pattern coverage-checking to `warn'.

location is the location to use for error-reporting in case of failure to perform with_warn_on_coverage_error ~location.

val with_coverage_checking : state -> location:Beluga_syntax.Syncom.Location.t -> (state -> 'a) -> 'a

with_coverage_checking state ~location enables coverage-checking.

location is the location to use for error-reporting in case of failure to perform with_coverage_checking ~location.

val set_name_generation_bases : state -> location:Beluga_syntax.Syncom.Location.t -> meta_variable_base:Beluga_syntax.Syncom.Identifier.t -> ?computation_variable_base:Beluga_syntax.Syncom.Identifier.t -> Beluga_syntax.Syncom.Qualified_identifier.t -> Stdlib.Unit.t

set_name_generation_bases state ~location ~meta_variable_base ?computation_variable_base constant sets the naming convention for name generation of meta-level and computation-level variables using meta_variable_base and computation_variable_base respectively for a type-level constant constant.

location is the location to use for error-reporting in case of failure to perform set_name_generation_bases ~location ~meta_variable_base ?computation_variable_base constant.

val set_default_associativity : state -> ?location:Beluga_syntax.Syncom.Location.t -> Beluga_syntax.Syncom.Associativity.t -> Stdlib.Unit.t

val get_default_associativity : state -> Beluga_syntax.Syncom.Associativity.t

val set_default_precedence : state -> ?location:Beluga_syntax.Syncom.Location.t -> Support.Int.t -> Stdlib.Unit.t

val get_default_precedence : state -> Support.Int.t

val add_prefix_notation : state -> ?location:Beluga_syntax.Syncom.Location.t -> ?precedence:Support.Int.t -> Beluga_syntax.Syncom.Qualified_identifier.t -> Stdlib.Unit.t

val add_infix_notation : state -> ?location:Beluga_syntax.Syncom.Location.t -> ?precedence:Support.Int.t -> ?associativity:Beluga_syntax.Syncom.Associativity.t -> Beluga_syntax.Syncom.Qualified_identifier.t -> Stdlib.Unit.t

val add_postfix_notation : state -> ?location:Beluga_syntax.Syncom.Location.t -> ?precedence:Support.Int.t -> Beluga_syntax.Syncom.Qualified_identifier.t -> Stdlib.Unit.t

val add_postponed_prefix_notation : state -> ?location:Beluga_syntax.Syncom.Location.t -> ?precedence:Support.Int.t -> Beluga_syntax.Syncom.Qualified_identifier.t -> Stdlib.Unit.t

add_postponed_prefix_notation state ?location ?precedence identifier adds a postponed prefix notation for identifier. If precedence = Option.None, then get_default_precedence is used instead.

This notation is postponed, meaning that it only applies once apply_postponed_fixity_pragmas is called.

val add_postponed_infix_notation : state -> ?location:Beluga_syntax.Syncom.Location.t -> ?precedence:Support.Int.t -> ?associativity:Beluga_syntax.Syncom.Associativity.t -> Beluga_syntax.Syncom.Qualified_identifier.t -> Stdlib.Unit.t

add_postponed_infix_notation state ?location ?precedence ?associativity identifier adds a postponed infix notation for identifier. If precedence = Option.None, then get_default_precedence is used instead. Likewise, if associativity = Option.None, then get_default_associativity is used instead.

This notation is postponed, meaning that it only applies once apply_postponed_fixity_pragmas is called.

val add_postponed_postfix_notation : state -> ?location:Beluga_syntax.Syncom.Location.t -> ?precedence:Support.Int.t -> Beluga_syntax.Syncom.Qualified_identifier.t -> Stdlib.Unit.t

add_postponed_postfix_notation state ?location ?precedence identifier adds a postponed postfix notation for identifier. If precedence = Option.None, then get_default_precedence is used instead.

This notation is postponed, meaning that it only applies once apply_postponed_fixity_pragmas is called.

val apply_postponed_fixity_pragmas : state -> unit

apply_postponed_fixity_pragmas state adds in scope the postponed prefix, infix and postfix fixity pragmas. This function should be called only when the targets of those postponed pragmas are in scope. That is, postponed fixity pragmas are applied after the subsequent declaration is added, or after a group of mutually recursive declarations are added.

val apply_postponed_fixity_pragmas_for_constant : state -> Beluga_syntax.Syncom.Identifier.t -> unit

apply_postponed_fixity_pragmas_for_constant state identifier adds in scope the postponed fixity pragmas for the constant having identifier identifier.

val open_module : state -> ?location:Beluga_syntax.Syncom.Location.t -> Beluga_syntax.Syncom.Qualified_identifier.t -> Stdlib.Unit.t

val add_module_abbreviation : state -> ?location:Beluga_syntax.Syncom.Location.t -> Beluga_syntax.Syncom.Qualified_identifier.t -> abbreviation:Beluga_syntax.Syncom.Identifier.t -> Stdlib.Unit.t

val with_checkpoint : state -> (state -> 'a) -> 'a

with_checkpoint state m is the result of running m with state. Note that the state is rolled back entirely.

ID Allocation

val next_module_id : state -> Beluga_syntax.Syncom.Id.module_id

Indexing

val index_lf_kind : state -> Synext.lf_kind -> Synapx.LF.kind

index_lf_kind state kind is kind indexed with respect to state. Free variables in kind are discarded.

val index_lf_typ : state -> Synext.lf_typ -> Synapx.LF.typ

val index_schema : state -> Synext.schema -> Synapx.LF.schema

val index_comp_kind : state -> Synext.comp_kind -> Synapx.Comp.kind

val index_comp_typ : state -> Synext.comp_typ -> Synapx.Comp.typ

val index_comp_expression : state -> Synext.comp_expression -> Synapx.Comp.exp

val index_comp_typedef : state -> Synext.comp_typ -> Synext.comp_kind -> Synapx.Comp.typ * Synapx.Comp.kind

val index_comp_theorem : state -> Synext.comp_expression -> Synapx.Comp.thm

val index_harpoon_proof : state -> Synext.harpoon_proof -> Synapx.Comp.thm

Adding a constant or module introduces an identifier in the state. Whenever an extensible declaration is shadowed, it is frozen.

val add_lf_type_constant : state -> ?location:Beluga_syntax.Syncom.Location.t -> Beluga_syntax.Syncom.Identifier.t -> Beluga_syntax.Syncom.Id.cid_typ -> Stdlib.Unit.t

add_lf_type_constant state ?location identifier id adds the LF type constant having identifier identifier, ID cid and binding site location to the state. If location = Option.None, then the identifier's location is used as binding site.

val add_lf_term_constant : state -> ?location:Beluga_syntax.Syncom.Location.t -> Beluga_syntax.Syncom.Identifier.t -> Beluga_syntax.Syncom.Id.cid_term -> Stdlib.Unit.t

val add_schema_constant : state -> ?location:Beluga_syntax.Syncom.Location.t -> Beluga_syntax.Syncom.Identifier.t -> Beluga_syntax.Syncom.Id.cid_schema -> Stdlib.Unit.t

val add_prog : state -> ?location:Beluga_syntax.Syncom.Location.t -> Beluga_syntax.Syncom.Identifier.t -> Beluga_syntax.Syncom.Id.cid_prog -> Stdlib.Unit.t

val add_comp_val : state -> ?location:Beluga_syntax.Syncom.Location.t -> Beluga_syntax.Syncom.Identifier.t -> Beluga_syntax.Syncom.Id.cid_prog -> Stdlib.Unit.t

val add_comp_typedef : state -> ?location:Beluga_syntax.Syncom.Location.t -> Beluga_syntax.Syncom.Identifier.t -> Beluga_syntax.Syncom.Id.cid_comp_typdef -> Stdlib.Unit.t

val add_comp_inductive_type_constant : state -> ?location:Beluga_syntax.Syncom.Location.t -> Beluga_syntax.Syncom.Identifier.t -> Beluga_syntax.Syncom.Id.cid_comp_typ -> Stdlib.Unit.t

val add_comp_stratified_type_constant : state -> ?location:Beluga_syntax.Syncom.Location.t -> Beluga_syntax.Syncom.Identifier.t -> Beluga_syntax.Syncom.Id.cid_comp_typ -> Stdlib.Unit.t

val add_comp_cotype_constant : state -> ?location:Beluga_syntax.Syncom.Location.t -> Beluga_syntax.Syncom.Identifier.t -> Beluga_syntax.Syncom.Id.cid_comp_cotyp -> Stdlib.Unit.t

val add_comp_constructor : state -> ?location:Beluga_syntax.Syncom.Location.t -> Beluga_syntax.Syncom.Identifier.t -> Beluga_syntax.Syncom.Id.cid_comp_const -> Stdlib.Unit.t

val add_comp_destructor : state -> ?location:Beluga_syntax.Syncom.Location.t -> Beluga_syntax.Syncom.Identifier.t -> Beluga_syntax.Syncom.Id.cid_comp_dest -> Stdlib.Unit.t

val add_module : state -> ?location:Beluga_syntax.Syncom.Location.t -> Beluga_syntax.Syncom.Identifier.t -> Beluga_syntax.Syncom.Id.module_id -> (state -> 'a) -> 'a

add_module state ?location identifier cid m adds the module having identifier identifier and binding site location to the state. The action m is executed in the module's state. m is typically the reconstruction of the module's entries.

val freeze_all_unfrozen_declarations : state -> Stdlib.Unit.t

freeze_all_unfrozen_declarations state freezes all the unfrozen declarations. Unfrozen declarations are LF type families in oldstyle syntax, and to which new LF term-level constants may be added.

val create_initial_state : Index_state.state -> state

val clear_state : clear_index_state:(Index_state.state -> unit) -> state -> unit