Module Beluga.ScopeCheck

type error =

| FixedFixed of string * Beluga_syntax.Common.Location.t * Id.cid_prog

| FloatingFloating of Id.cid_prog * Id.cid_prog

exception Error of Beluga_syntax.Common.Location.t * error

val throw : Beluga_syntax.Common.Location.t -> error -> 'a

val thm_thm : Beluga_syntax.Common.Location.t -> Id.cid_prog -> Id.cid_prog option -> unit

Perform late scopechecking for a theorem-theorem interaction. thm_thm loc cid mcid checks that the theorem identified by `cid` is in scope for the theorem specified by `mcid`, if any.

val other_thm : Beluga_syntax.Common.Location.t -> string -> Decl.t -> Beluga_syntax.Common.Location.t -> Id.cid_prog option -> unit

Performs late scopechecking for a other-theorem interaction. other_thm loc name d loc' mcid checks that the declaration `d` located at `loc'` is in scope for the theorem specified by `mcid`, if any.

val format_error : Support.Format.formatter -> error -> unit