Module Store.CVar

Persistent stack-based store for contextual variables.

Used for indexing.

type cvar = Name.t

type entry = {

name : cvar;

plicity : Beluga_syntax.Common.Plicity.t;

}

val mk_entry : cvar -> Beluga_syntax.Common.Plicity.t -> entry

type t

val empty : t

val extend : t -> entry -> t

val get : t -> Id.var -> entry

val index_of_name : t -> cvar -> Beluga_syntax.Common.Plicity.t * Id.offset

index_of_name scope name is (plicity, offset) where offset is the de Bruijn index of name in scope, and plicity is whether that variable was explicitly declared or not.

raises Not_found
if no such variable is in scope.

val append : t -> t -> t

val length : t -> int

val of_mctx : ((Beluga_syntax.Common.Plicity.t * Beluga_syntax.Common.Inductivity.t) -> Beluga_syntax.Common.Plicity.t) -> Beluga.Syntax.Int.LF.mctx -> t

Erases the context down to a list of names using the given function to interpret the plicity of declarations.

val to_string : t -> string

val of_list : (Name.t * Beluga_syntax.Common.Plicity.t) list -> t