Module Cid.Comp

val add_mutual_group : Beluga_syntax.Synint.Comp.total_dec list -> Beluga_syntax.Id.cid_mutual_group

module Entry : sig ... end

type entry = Entry.t

val name : Beluga_syntax.Id.cid_prog -> Beluga_syntax.Name.t

Gets the Name.t of the given theorem ID.

val lookup_mutual_group : Beluga_syntax.Id.cid_mutual_group -> Beluga_syntax.Synint.Comp.total_dec list

Looks up the total declarations for the given mutual group.

val mutual_group : Beluga_syntax.Id.cid_prog -> Beluga_syntax.Id.cid_mutual_group

Gets the mutual group ID for a given function reference.

val total_decs : Beluga_syntax.Id.cid_prog -> Beluga_syntax.Synint.Comp.total_dec list

Looks up the total declarations for the mutual group of the given function.

val get_total_decl : Beluga_syntax.Id.cid_prog -> Beluga_syntax.Synint.Comp.total_dec

Looks up the totality declaration for the given theorem.

val mk_entry : Beluga_syntax.Name.t -> Beluga_syntax.Synint.Comp.typ -> int -> Beluga_syntax.Id.cid_mutual_group -> Beluga_syntax.Synint.Comp.value option -> entry

val add : (Beluga_syntax.Id.cid_prog -> entry) -> Beluga_syntax.Id.cid_prog

If the value we store in the entry is a recursive value, it itself needs the cid_prog that we are creating to store this entry. Therefore, unlike 'add' functions in other modules, this 'add' function expects a function to which it will provide the cid_prog it generated to store the entry, thus tying the recursive knot.

val get : Beluga_syntax.Id.cid_prog -> entry

val current_entries : unit -> (Beluga_syntax.Id.cid_prog * entry) list

val explicit_arguments : entry -> int

val fixed_name_of : Beluga_syntax.Id.cid_prog -> Beluga_syntax.Name.t

val clear : unit -> unit

val set_prog : Beluga_syntax.Id.cid_prog -> (Beluga_syntax.Synint.Comp.value option -> Beluga_syntax.Synint.Comp.value option) -> unit

Update the associated program of an existing entry. This is notably used for two-step elaboration of functions:

a function is added to the store with its type *before* its bodies is elaborated

indexing can find the cid of the function immediately from the store.

when the function body is elaborated, the store entry is updated to hold the body as well.