Module Beluga.Total

Checking termination of a function

module Unify = Unify.StdTrail

module P = Prettyint.DefaultPrinter

module R = Store.Cid.DefaultRenderer

val dprintf : Support.Debug.Fmt.fmt Support.Debug.io Support.Debug.io

val dprint : (unit -> string) Support.Debug.io

val dprnt : string Support.Debug.io

type error =

| NoPositiveCheck of string

| NoStratifyCheck of string

| NoStratifyOrPositiveCheck of string

| WrongArgNum of Beluga_syntax.Id.cid_comp_typ * int

| RecCallIncompatible of Beluga_syntax.Int.LF.mctx * Beluga_syntax.Int.Comp.ih_arg * Beluga_syntax.Int.Comp.ih_decl

| NotImplemented of string

| TooManyArg

exception E of Beluga_syntax.Location.t * error

val throw : Beluga_syntax.Location.t -> error -> 'a

val error_printer : error -> Support.Format.formatter -> unit

val print_str : (unit -> string) -> unit

exception Not_compatible

exception CtxNot_compatible

val enabled : bool Stdlib.ref

type rec_arg =

| M of Beluga_syntax.Int.Comp.meta_obj

| V of Beluga_syntax.Int.Comp.exp

val shiftArg : int -> Beluga_syntax.Int.Comp.ih_arg -> Beluga_syntax.Int.Comp.ih_arg

val shiftIH : int -> Beluga_syntax.Int.Comp.ih_decl Beluga_syntax.Int.LF.ctx -> Beluga_syntax.Int.Comp.ih_decl Beluga_syntax.Int.LF.ctx

val shift : Beluga_syntax.Int.Comp.ih_decl Beluga_syntax.Int.LF.ctx -> Beluga_syntax.Int.Comp.ih_decl Beluga_syntax.Int.LF.ctx

val sub_smaller : ('a * int) -> Beluga_syntax.Int.LF.sub -> bool

val smaller_meta_obj : Beluga_syntax.Int.LF.mfront -> bool

val mark_gctx : Beluga_syntax.Int.Comp.ctyp_decl Beluga_syntax.Int.LF.ctx -> Beluga_syntax.Int.Comp.ctyp_decl Beluga_syntax.Int.LF.ctx

val struct_smaller : Beluga_syntax.Int.Comp.pattern -> bool

val order_to_string : int Beluga_syntax.Int.Comp.generic_order option -> string

val is_valid_args : Beluga_syntax.Int.Comp.typ -> int -> bool

Determine whether the arguments given in a totality declaration are valid. n: number of args tau: type of the recursive function

val get_order : Beluga_syntax.Int.Comp.total_dec list -> (Syncom.Name.t * int list option * (Beluga_syntax.Int.Comp.typ * Beluga_syntax.Synint.LF.msub)) list

Converts the list of totality declarations into an induction ordering list

val lookup_dec : Beluga_syntax.Name.t -> Beluga_syntax.Int.Comp.total_dec list -> Beluga_syntax.Int.Comp.total_dec option

Looks up a declaration based on a function name

val get_order_for : Beluga_syntax.Int.Comp.total_dec list -> Beluga_syntax.Name.t -> int list option

Gets the induction order for the function with name `f` Returns None if the

val gen_var' : Beluga_syntax.Location.t -> Beluga_syntax.Synint.LF.mctx -> (Beluga_syntax.Name.t * Beluga_syntax.Int.LF.ctyp) -> (Beluga_syntax.Location.t * Beluga_syntax.Int.LF.mfront) * Beluga_syntax.Int.LF.mfront

val gen_var : Beluga_syntax.Location.t -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Int.LF.ctyp_decl -> ((Beluga_syntax.Location.t * Beluga_syntax.Int.LF.mfront) * Beluga_syntax.Int.LF.mfront) * Beluga_syntax.Int.LF.ctyp

val gen_meta_obj : (Syncom.Id.offset * Beluga_syntax.Int.LF.sub) -> Beluga_syntax.Int.LF.ctyp -> Beluga_syntax.Location.t * Beluga_syntax.Int.LF.mfront

val uninstantiated_arg : Beluga_syntax.Synint.Comp.meta_obj -> bool

val generalize : Beluga_syntax.Int.Comp.ih_arg list -> Beluga_syntax.Int.Comp.ih_arg list

val valid_args : Beluga_syntax.Int.Comp.ih_arg list -> bool

val genTypeAgainstSchema : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> 'a -> Beluga_syntax.Synint.LF.typ -> Beluga_syntax.Int.LF.schema -> ('b * Beluga_syntax.Synint.LF.typ) option

val gen_ctx : 'a -> 'b -> (Beluga_syntax.Id.cid_schema * 'c) -> 'd * Beluga_syntax.Synint.LF.sub

val gen_arg : Beluga_syntax.Synint.LF.mctx -> (Syncom.Id.offset * Beluga_syntax.Synint.Comp.meta_typ) -> Beluga_syntax.Int.LF.ctyp -> Beluga_syntax.Location.t * Beluga_syntax.Int.LF.mfront

val rec_spine : Beluga_syntax.Synint.LF.mctx -> (Syncom.Id.offset * Beluga_syntax.Synint.Comp.meta_typ) -> (int * (Beluga_syntax.Synint.Comp.typ * Beluga_syntax.Synint.LF.msub)) -> Beluga_syntax.Int.Comp.ih_arg list * Beluga_syntax.Synint.Comp.typ

val rec_spine' : Beluga_syntax.Synint.LF.mctx -> (Syncom.Id.offset * (Beluga_syntax.Synint.Comp.typ * Beluga_syntax.Synint.LF.msub)) -> (int * (Beluga_syntax.Synint.Comp.typ * Beluga_syntax.Synint.LF.msub)) -> Beluga_syntax.Int.Comp.ih_arg list * Beluga_syntax.Synint.Comp.typ

val gen_rec_calls : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Int.Comp.ih_decl Beluga_syntax.Int.LF.ctx -> (Beluga_syntax.Int.LF.ctyp_decl Beluga_syntax.Int.LF.ctx * int) -> Beluga_syntax.Int.Comp.total_dec list -> Beluga_syntax.Int.Comp.ih_decl Beluga_syntax.Int.LF.ctx

val get_return_type : Beluga_syntax.Synint.LF.mctx -> 'a -> (Beluga_syntax.Int.Comp.typ * Beluga_syntax.Synint.LF.msub) -> 'b * (Beluga_syntax.Int.Comp.typ * Beluga_syntax.Synint.LF.msub)

val gen_rec_calls' : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.Comp.gctx -> Beluga_syntax.Int.Comp.ih_decl Beluga_syntax.Int.LF.ctx -> (Beluga_syntax.Int.Comp.ctyp_decl Beluga_syntax.Int.LF.ctx * Syncom.Id.offset) -> Beluga_syntax.Int.Comp.total_dec list -> Beluga_syntax.Int.Comp.ih_decl Beluga_syntax.Int.LF.ctx

val wf_rec_calls : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.Comp.gctx -> Beluga_syntax.Int.Comp.total_dec Support.List.t -> Beluga_syntax.Int.Comp.ih_decl Beluga_syntax.Int.LF.ctx

Computes the well-founded recursive calls for the given group of mutual functions on the given contexts. If the list of functions is empty, then an empty context is computed.

exception WkViolation

val wkSub : Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.sub

val pos : Beluga_syntax.Int.LF.dctx -> Beluga_syntax.Synint.LF.typ -> int -> int option

val weakSub : 'a -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.sub

val convDCtxMod : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.sub option

val convDCtxGen' : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Int.LF.dctx -> Beluga_syntax.Int.LF.dctx -> Beluga_syntax.Int.LF.sub option

val convDCtxGen : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Int.LF.sub option

val prefixDCtx' : 'a -> Beluga_syntax.Int.LF.dctx -> Beluga_syntax.Int.LF.dctx -> int option

val prefixDCtx : 'a -> Beluga_syntax.Int.LF.dctx -> Beluga_syntax.Int.LF.dctx -> int option

val prefix_hat : (Beluga_syntax.Int.LF.ctx_var option * int) -> (Beluga_syntax.Int.LF.ctx_var option * int) -> int option

val dot_k : Beluga_syntax.Synint.LF.sub -> int -> Beluga_syntax.Synint.LF.sub

val shiftMetaObj : Beluga_syntax.Synint.LF.mctx -> ((Syncom.Location.t * Beluga_syntax.Int.LF.mfront) * Beluga_syntax.Int.LF.ctyp) -> (Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.sub * Beluga_syntax.Synint.LF.dctx) -> (Syncom.Location.t * Beluga_syntax.Int.LF.mfront) * Beluga_syntax.Int.LF.ctyp

val shiftArgs : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Int.Comp.ih_arg list -> (Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.sub * Beluga_syntax.Synint.LF.dctx) -> Beluga_syntax.Int.Comp.ih_arg list

val filter : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.Comp.gctx -> Beluga_syntax.Int.Comp.ih_decl Beluga_syntax.Int.LF.ctx -> ('a * Beluga_syntax.Int.Comp.ih_arg) -> Beluga_syntax.Int.Comp.ih_decl Beluga_syntax.Int.LF.ctx

val annotate' : Beluga_syntax.Synint.Comp.typ -> int list -> Beluga_syntax.Synint.Comp.typ option

Adjusts the given type signature with annotations for the induction arguments according to the given order. Returns Option.None if the order does not match the type, i.e. goes out of bounds.

val strip : Beluga_syntax.Synint.Comp.typ -> Beluga_syntax.Synint.Comp.typ

Removes all TypInd marks in a computational type.

exception Unimplemented

val no_occurs : Beluga_syntax.Id.cid_comp_typ -> Beluga_syntax.Int.Comp.typ -> bool

val check_positive : Beluga_syntax.Id.cid_comp_typ -> Beluga_syntax.Int.Comp.typ -> bool

val positive : Beluga_syntax.Id.cid_comp_typ -> Beluga_syntax.Int.Comp.typ -> bool

val less_dctx : Beluga_syntax.Int.LF.dctx -> Beluga_syntax.Int.LF.dctx -> bool

val equal_meta_obj : 'a -> Beluga_syntax.Synint.Comp.meta_obj -> Beluga_syntax.Synint.Comp.meta_obj -> bool

val less_meta_obj : 'a -> Beluga_syntax.Synint.Comp.meta_obj -> Beluga_syntax.Synint.Comp.meta_obj -> bool

val leq_meta_obj : 'a -> Beluga_syntax.Synint.Comp.meta_obj -> Beluga_syntax.Synint.Comp.meta_obj -> bool

val find_meta_obj : Beluga_syntax.Int.Comp.meta_spine -> int -> Beluga_syntax.Int.Comp.meta_obj

val get_target : Synint.LF.ctyp_decl Beluga_syntax.Int.LF.ctx -> Beluga_syntax.Int.Comp.typ -> Synint.LF.ctyp_decl Beluga_syntax.Int.LF.ctx * Beluga_syntax.Int.Comp.meta_spine

val mS_size : Beluga_syntax.Int.Comp.meta_spine -> int

val compare : Beluga_syntax.Id.cid_comp_typ -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Int.Comp.typ -> Beluga_syntax.Synint.Comp.meta_obj -> int -> bool

val stratify : Beluga_syntax.Id.cid_comp_typ -> Beluga_syntax.Int.Comp.typ -> int -> bool

val stratifyAll : Beluga_syntax.Id.cid_comp_typ -> Beluga_syntax.Int.Comp.typ -> int

val is_comp_inductive : Beluga_syntax.Int.Comp.gctx -> Beluga_syntax.Int.Comp.exp -> bool

Decides whether a data object is something we're doing induction on, i.e. it's a computational variable whose type is TypInd or with a WF flag

This is similar to the logic used in check.ml to determine the kind of a branch: IndDataObj or IndIndexObj.

val is_meta_inductive : Beluga_syntax.Int.LF.mctx -> Beluga_syntax.Int.LF.mfront -> bool

Decides whether an index object is something we're doing induction on, i.e. it's a metavariable with the Inductive flag set when we look it up in the meta-context.

val is_inductive_split : Beluga_syntax.Int.LF.mctx -> Beluga_syntax.Int.Comp.gctx -> Beluga_syntax.Int.Comp.exp -> bool

Checks if the scrutinee of a case is on an inductive computational variable or an inductive meta-variable.

val requires_checking : Beluga_syntax.Name.t -> Beluga_syntax.Int.Comp.total_dec list -> bool

Decides, in the context of a given list of totality declarations, whether the given function (name) requires totality checking. A function requires totality checking only if its totality declaration in the given list, if any, is an `inductive declaraction.

val annotate : Beluga_syntax.Location.t -> [> `inductive of Beluga.Order.I.order ] -> Beluga_syntax.Synint.Comp.typ -> Beluga_syntax.Synint.Comp.typ

Applies the given induction order to produce an annotated type.

val select_ihs : Beluga_syntax.Name.t -> Beluga_syntax.Int.Comp.ih_decl Beluga_syntax.Int.LF.ctx -> Beluga_syntax.Int.Comp.ih_decl Beluga_syntax.Int.LF.ctx

Filters an ihctx to contain only those IHs for the given function name.

val drop_args : int -> Beluga_syntax.Int.Comp.ih_decl Beluga_syntax.Synint.LF.ctx -> Beluga_syntax.Int.Comp.ih_decl Beluga_syntax.Synint.LF.ctx

Drops n leading arguments from all recursive calls in a given ihctx.

val drop_arg : Beluga_syntax.Int.Comp.ih_decl Beluga_syntax.Synint.LF.ctx -> Beluga_syntax.Int.Comp.ih_decl Beluga_syntax.Synint.LF.ctx

Drops one argument from all recursive calls in a given ihctx.