Module Beluga.Whnf

author Brigitte Pientka

modified: Joshua Dunfield

val whnf : Beluga_syntax.Synint.LF.nclo -> Beluga_syntax.Synint.LF.nclo

val whnfTyp : Beluga_syntax.Synint.LF.tclo -> Beluga_syntax.Synint.LF.tclo

val norm : Beluga_syntax.Synint.LF.nclo -> Beluga_syntax.Synint.LF.normal

val normKind : (Beluga_syntax.Synint.LF.kind * Beluga_syntax.Synint.LF.sub) -> Beluga_syntax.Synint.LF.kind

val normTyp : Beluga_syntax.Synint.LF.tclo -> Beluga_syntax.Synint.LF.typ

val normFt' : (Beluga_syntax.Synint.LF.front * Beluga_syntax.Synint.LF.sub) -> Beluga_syntax.Synint.LF.front

val normClTyp : (Beluga_syntax.Synint.LF.cltyp * Beluga_syntax.Synint.LF.sub) -> Beluga_syntax.Synint.LF.cltyp

val normTypRec : Beluga_syntax.Synint.LF.trec_clo -> Beluga_syntax.Synint.LF.typ_rec

val normSub : Beluga_syntax.Synint.LF.sub -> Beluga_syntax.Synint.LF.sub

val normSpine : Beluga_syntax.Synint.LF.sclo -> Beluga_syntax.Synint.LF.spine

val normDCtx : Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.dctx

val normMCtx : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.mctx

val reduce : Beluga_syntax.Synint.LF.nclo -> Beluga_syntax.Synint.LF.spine -> Beluga_syntax.Synint.LF.normal

val whnfRedex : (Beluga_syntax.Synint.LF.nclo * Beluga_syntax.Synint.LF.sclo) -> Beluga_syntax.Synint.LF.nclo

val conv : Beluga_syntax.Synint.LF.nclo -> Beluga_syntax.Synint.LF.nclo -> bool

val convHead : (Beluga_syntax.Synint.LF.head * Beluga_syntax.Synint.LF.sub) -> (Beluga_syntax.Synint.LF.head * Beluga_syntax.Synint.LF.sub) -> bool

val convTyp : Beluga_syntax.Synint.LF.tclo -> Beluga_syntax.Synint.LF.tclo -> bool

val convTypRec : Beluga_syntax.Synint.LF.trec_clo -> Beluga_syntax.Synint.LF.trec_clo -> bool

val convSchElem : Beluga_syntax.Synint.LF.sch_elem -> Beluga_syntax.Synint.LF.sch_elem -> bool

val prefixSchElem : Beluga_syntax.Synint.LF.sch_elem -> Beluga_syntax.Synint.LF.sch_elem -> bool

val convSub : Beluga_syntax.Synint.LF.sub -> Beluga_syntax.Synint.LF.sub -> bool

val convITerm : Beluga_syntax.Synint.LF.iterm -> Beluga_syntax.Synint.LF.iterm -> bool

val convMSub : Beluga_syntax.Synint.LF.msub -> Beluga_syntax.Synint.LF.msub -> bool

val convCtxVar : Beluga_syntax.Synint.LF.ctx_var -> Beluga_syntax.Synint.LF.ctx_var -> bool

val convDCtx : Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.dctx -> bool

val convDCtxHat : Beluga_syntax.Synint.LF.dctx_hat -> Beluga_syntax.Synint.LF.dctx_hat -> bool

val newMMVar' : Beluga_syntax.Name.t option -> (Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.LF.ctyp) -> Beluga_syntax.Plicity.t -> Beluga_syntax.Inductivity.t -> Beluga_syntax.Synint.LF.mm_var

val newMMVar : Beluga_syntax.Name.t option -> (Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.typ) -> Beluga_syntax.Plicity.t -> Beluga_syntax.Inductivity.t -> Beluga_syntax.Synint.LF.mm_var

val newMPVar : Beluga_syntax.Name.t option -> (Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.typ) -> Beluga_syntax.Plicity.t -> Beluga_syntax.Inductivity.t -> Beluga_syntax.Synint.LF.mm_var

val newMSVar : Beluga_syntax.Name.t option -> (Beluga_syntax.Synint.LF.mctx * Beluga_syntax.Synint.LF.svar_class * Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.dctx) -> Beluga_syntax.Plicity.t -> Beluga_syntax.Inductivity.t -> Beluga_syntax.Synint.LF.mm_var

val newMVar : Beluga_syntax.Name.t option -> (Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.typ) -> Beluga_syntax.Plicity.t -> Beluga_syntax.Inductivity.t -> Beluga_syntax.Synint.LF.cvar

val newCVar : Beluga_syntax.Name.t option -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Id.cid_schema option -> Beluga_syntax.Plicity.t -> Beluga_syntax.Inductivity.t -> Beluga_syntax.Synint.LF.ctx_var

val raiseType : Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.typ -> Beluga_syntax.Synint.LF.typ

val etaExpandMV : Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.tclo -> Beluga_syntax.Name.t -> Beluga_syntax.Synint.LF.sub -> Beluga_syntax.Plicity.t -> Beluga_syntax.Inductivity.t -> Beluga_syntax.Synint.LF.normal

val etaExpandMMV : Beluga_syntax.Location.t -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.tclo -> Beluga_syntax.Name.t -> Beluga_syntax.Synint.LF.sub -> Beluga_syntax.Plicity.t -> Beluga_syntax.Inductivity.t -> Beluga_syntax.Synint.LF.normal

exception Fmvar_not_found

exception FreeMVar of Beluga_syntax.Synint.LF.head

exception NonInvertible

exception InvalidLFHole of Beluga_syntax.Location.t

val newMTypName : Beluga_syntax.Synint.LF.ctyp -> Beluga_syntax.Name.t

val m_id : Beluga_syntax.Synint.LF.msub

val mvar_dot1 : Beluga_syntax.Synint.LF.msub -> Beluga_syntax.Synint.LF.msub

val pvar_dot1 : Beluga_syntax.Synint.LF.msub -> Beluga_syntax.Synint.LF.msub

val mvar_dot : Beluga_syntax.Synint.LF.msub -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.msub

val mcomp : Beluga_syntax.Synint.LF.msub -> Beluga_syntax.Synint.LF.msub -> Beluga_syntax.Synint.LF.msub

mcomp t1 t2 = t' Eagerly composes the modal substitutions t1 and t2.

If cD_2 |- t1 : cD_1 and cD_3 |- t2 : cD_2 then cD_3 |- t' : cD_1 where t' = mcomp t1 t2

For example, suppose cD |- tau <= type for some tau cD' |- t' : cD cD'' |- t'' : cD'

cnormCTyp (tau, mcomp t' t'') = cnormCTyp (cnormCTyp (tau, t'), t'')

That is, applying a composition of substitutions is equivalent to applying the composed substitutions from *left-to-right*.

val mcomp' : Beluga_syntax.Synint.LF.msub -> Beluga_syntax.Synint.LF.msub -> Beluga_syntax.Synint.LF.msub

Flipped version of mcomp which is more useful as a higher-order function.

If cD_2 |- t1 : cD_1 and cD_3 |- t2 : cD_2 then cD_3 |- t' : cD_1 where t' = mcomp' t2 t1

val m_invert : Beluga_syntax.Synint.LF.msub -> Beluga_syntax.Synint.LF.msub

val mctxLookup : Beluga_syntax.Synint.LF.mctx -> int -> Beluga_syntax.Name.t * Beluga_syntax.Synint.LF.ctyp

val mctxLookupDep : Beluga_syntax.Synint.LF.mctx -> int -> Beluga_syntax.Name.t * Beluga_syntax.Synint.LF.ctyp * Beluga_syntax.Plicity.t * Beluga_syntax.Inductivity.t

val mctxMDec : Beluga_syntax.Synint.LF.mctx -> int -> Beluga_syntax.Name.t * Beluga_syntax.Synint.LF.typ * Beluga_syntax.Synint.LF.dctx

val mctxPDec : Beluga_syntax.Synint.LF.mctx -> int -> Beluga_syntax.Name.t * Beluga_syntax.Synint.LF.typ * Beluga_syntax.Synint.LF.dctx

val mctxSDec : Beluga_syntax.Synint.LF.mctx -> int -> Beluga_syntax.Name.t * Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.svar_class * Beluga_syntax.Synint.LF.dctx

val mctxCDec : Beluga_syntax.Synint.LF.mctx -> int -> Beluga_syntax.Name.t * Beluga_syntax.Id.cid_schema

val mctxMVarPos : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Name.t -> Beluga_syntax.Id.offset * Beluga_syntax.Synint.LF.ctyp

val cnorm : (Beluga_syntax.Synint.LF.normal * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.LF.normal

val cnormHead : (Beluga_syntax.Synint.LF.head * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.LF.head

val cnormHead' : (Beluga_syntax.Synint.LF.head * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.LF.front

val cnormSpine : (Beluga_syntax.Synint.LF.spine * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.LF.spine

val cnormSub : (Beluga_syntax.Synint.LF.sub * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.LF.sub

val cnormTyp : (Beluga_syntax.Synint.LF.typ * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.LF.typ

val cnormTypRec : (Beluga_syntax.Synint.LF.typ_rec * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.LF.typ_rec

val cnormDCtx : (Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.LF.dctx

val cnormMTyp : (Beluga_syntax.Synint.LF.ctyp * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.LF.ctyp

val cnormClTyp : (Beluga_syntax.Synint.LF.cltyp * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.LF.cltyp

val cnorm_psihat : Beluga_syntax.Synint.LF.dctx_hat -> Beluga_syntax.Synint.LF.msub -> Beluga_syntax.Synint.LF.dctx_hat

val cnormGCtx : (Beluga_syntax.Synint.Comp.gctx * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.Comp.gctx

val cnormIHCtx : (Beluga_syntax.Synint.Comp.ihctx * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.Comp.ihctx

val cnormPattern : (Beluga_syntax.Synint.Comp.pattern * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.Comp.pattern

val cnormPatSpine : (Beluga_syntax.Synint.Comp.pattern_spine * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.Comp.pattern_spine

val cnormMetaObj : (Beluga_syntax.Synint.Comp.meta_obj * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.Comp.meta_obj

val cnormClObj : Beluga_syntax.Synint.LF.clobj -> Beluga_syntax.Synint.LF.msub -> Beluga_syntax.Synint.LF.clobj

val cnormMFt : Beluga_syntax.Synint.LF.mfront -> Beluga_syntax.Synint.LF.msub -> Beluga_syntax.Synint.LF.mfront

val cnormMSub : Beluga_syntax.Synint.LF.msub -> Beluga_syntax.Synint.LF.msub

val cnormCCDecl : (Beluga_syntax.Synint.Comp.ctyp_decl * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.Comp.ctyp_decl

val cnormCKind : (Beluga_syntax.Synint.Comp.kind * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.Comp.kind

val cnormCTyp : (Beluga_syntax.Synint.Comp.typ * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.Comp.typ

val cnormCDecl : (Beluga_syntax.Synint.LF.ctyp_decl * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.LF.ctyp_decl

val cwhnfCTyp : (Beluga_syntax.Synint.Comp.typ * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.Comp.typ * Beluga_syntax.Synint.LF.msub

val cwhnfCtx : (Beluga_syntax.Synint.Comp.gctx * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.Comp.gctx

val cnormExp : (Beluga_syntax.Synint.Comp.exp * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.Comp.exp

val cnormThm : (Beluga_syntax.Synint.Comp.thm * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.Comp.thm

val normGCtx : Beluga_syntax.Synint.Comp.gctx -> Beluga_syntax.Synint.Comp.gctx

val normIHCtx : Beluga_syntax.Synint.Comp.ihctx -> Beluga_syntax.Synint.Comp.ihctx

val normCTyp : Beluga_syntax.Synint.Comp.typ -> Beluga_syntax.Synint.Comp.typ

val convMTyp : Beluga_syntax.Synint.LF.ctyp -> Beluga_syntax.Synint.LF.ctyp -> bool

val convCTypDecl : Beluga_syntax.Synint.LF.ctyp_decl -> Beluga_syntax.Synint.LF.ctyp_decl -> bool

val convCTyp : (Beluga_syntax.Synint.Comp.typ * Beluga_syntax.Synint.LF.msub) -> (Beluga_syntax.Synint.Comp.typ * Beluga_syntax.Synint.LF.msub) -> bool

val convMetaObj : Beluga_syntax.Synint.Comp.meta_obj -> Beluga_syntax.Synint.Comp.meta_obj -> bool

val convCompCTypDecl : Beluga_syntax.Synint.Comp.ctyp_decl -> Beluga_syntax.Synint.Comp.ctyp_decl -> bool

val closed : Beluga_syntax.Synint.LF.nclo -> bool

val closedTyp : Beluga_syntax.Synint.LF.tclo -> bool

val closedDCtx : Beluga_syntax.Synint.LF.dctx -> bool

val closedCTyp : Beluga_syntax.Synint.Comp.typ -> bool

val closedGCtx : Beluga_syntax.Synint.Comp.gctx -> bool

val closedMetaObj : Beluga_syntax.Synint.Comp.meta_obj -> bool

val closedExp : Beluga_syntax.Synint.Comp.exp -> bool

val constraints_solved : Beluga_syntax.Synint.LF.cnstr list -> bool

val mctx_to_list_shifted : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.ctyp_decl list

Convert a meta-context to a list of declarations in which the * types have been shifted to make sense in the whole context. * * This would go in Context, but we need to invoke Whnf to perform * the shifting, and Whnf depends on Context.

val collapse_sigma : Beluga_syntax.Synint.LF.typ_rec -> Beluga_syntax.Synint.LF.typ

Eliminates the level of indirection from having a Sigma with just one component.

val conv_subgoal_path : Beluga_syntax.Synint.Comp.SubgoalPath.t -> Beluga_syntax.Synint.Comp.SubgoalPath.t -> bool

val conv_subgoal_path_builder : Beluga_syntax.Synint.Comp.SubgoalPath.builder -> Beluga_syntax.Synint.Comp.SubgoalPath.builder -> bool

val convMCtx : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.mctx -> bool

Checks strong convertibility between two meta-contexts. We say _strong_ because reordering of entries is forbidden.

val convGCtx : (Beluga_syntax.Synint.Comp.gctx * Beluga_syntax.Synint.LF.msub) -> (Beluga_syntax.Synint.Comp.gctx * Beluga_syntax.Synint.LF.msub) -> bool

Checks strong convertibility between two computational contexts. We say _strong_ because reordering of entries is forbidden.

val lowerTyp : Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.tclo -> Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.tclo

lowerTyp cPsi (tA, s) = (cPsi', (tA', s')) where

cPsi |- s : cPhi

cPsi |- tAs <= type

cPsi' |- tA's' <= type

cPsi' is an extension of cPsi formed by shifting all Pi-type declarations in tA to cPsi.

val mmVarToClObj : Beluga_syntax.Location.t -> Beluga_syntax.Synint.LF.mm_var -> Beluga_syntax.Synint.LF.cltyp -> Beluga_syntax.Synint.LF.clobj

Converts an MMVar to a contextual object according to its contextual type.

val mmVarToMFront : Beluga_syntax.Location.t -> Beluga_syntax.Synint.LF.mm_var -> Beluga_syntax.Synint.Comp.meta_typ -> Beluga_syntax.Synint.LF.mfront

Converts an MMVar to a meta object according to its meta type.

val dotMMVar : Beluga_syntax.Location.t -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.msub -> (Beluga_syntax.Name.t * Beluga_syntax.Synint.LF.ctyp * Beluga_syntax.Plicity.t * Beluga_syntax.Inductivity.t) -> Beluga_syntax.Synint.Comp.meta_obj * Beluga_syntax.Synint.LF.msub

val extend_mctx : Beluga_syntax.Synint.LF.mctx -> (Beluga_syntax.Synint.LF.ctyp_decl * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.LF.mctx

val extend_gctx : Beluga_syntax.Synint.Comp.gctx -> (Beluga_syntax.Synint.Comp.ctyp_decl * Beluga_syntax.Synint.LF.msub) -> Beluga_syntax.Synint.Comp.gctx