Module Centiparsec.Shunting_yard

Implementation of Dijkstra's shunting yard algorithm for parsing expression operators using reverse Polish notation.

This is used for parsing terms having operators with user-defined precedences, fixities, and associativities.

module type ASSOCIATIVITY = sig ... end

module type FIXITY = sig ... end

module type OPERAND = sig ... end

module type OPERATOR = sig ... end

module type S = sig ... end

module Make (Associativity : ASSOCIATIVITY) (Fixity : FIXITY) (Operand : OPERAND) (Operator : OPERATOR with type associativity = Associativity.t and type fixity = Fixity.t and type operand = Operand.t) : S with type operand = Operand.t and type operator = Operator.t

Make (Associativity) (Fixity) (Operand) (Operator) is a module implementing Dijkstra's shunting yard algorithm for rewriting lists of Operand.t to an Operand.t with dynamic (or user-defined) operators.