Module Shunting_yard.Make

Make (Associativity) (Fixity) (Operand) (Operator) is a module implementing Dijkstra's shunting yard algorithm for rewriting lists of Operand.t to an Operand.t with dynamic (or user-defined) operators.

This implementation supports:

prefix operators of arbitrary arity

left-associative, right-associative and non-associative infix operators

postfix operators of arbitrary arity

Parameters

module Associativity : ASSOCIATIVITY

module Fixity : FIXITY

module Operand : OPERAND

module Operator : OPERATOR with type associativity = Associativity.t and type fixity = Fixity.t and type operand = Operand.t

Signature

type operand = Operand.t

type operator = Operator.t

type primitive

The type of expressions tagged as either an operand or an operator for rewriting with shunting_yard.

Constructors

val operand : operand -> primitive

val operator : operator -> primitive

Rewriting

exception Empty_expression

Empty_expression is raised from shunting_yard primitives if primitives = [].

exception Misplaced_operator of {

operator : operator;

operands : operand list;

}

Misplaced_operator { operator; operands } is raised from shunting_yard primitives if operator appears in an illegal position in primitives such that it would be rewritten with arguments arguments.

exception Ambiguous_operator_placement of {

left_operator : operator;

right_operator : operator;

}

Ambiguous_operator_placement { left_operator; right_operator } is raised from shunting_yard primitives if right_operator appears in an illegal position in primitives with respect to left_operator, such that left_operator and right_operator have the same precedence, and would be rewritten as being consecutive while one of left_operator and right_operator is left-associative with the other being right-associative.

For instance:

a -> b <- c is ambiguous if -> is right-associative and <- is left-associative, with -> and <- having the same precedence.

a <- b -> c is ambiguous if <- is left-associative and -> is right-associative, with <- and -> having the same precedence.

++ a + b is ambiguous if ++ is right-associative and + is left-associative, with ++ and + having the same precedence.

++ a ! is ambiguous if ++ is right-associative and ! is left-associative, with ++ and ! having the same precedence.

a ^ b ! is ambiguous if ^ is right-associative and ! is left-associative, with ^ and ! having the same precedence.

exception Consecutive_non_associative_operators of {

left_operator : operator;

right_operator : operator;

}

Consecutive_non_associative_operators { left_operator; right_operator } is raised from shunting_yard primitives if non-associative operator right_operator appears in an illegal position in primitives with respect to left_operator, such that left_operator = right_operator, and left_operator and right_operator would be rewritten as being consecutive, as in a left_operator b right_operator c.

exception Arity_mismatch of {

operator : operator;

operator_arity : int;

operands : operand list;

}

Arity_mismatch { operator; operator_arity; operands } is raised from shunting_yard primitives if operator appears in an illegal position in primitives such that it is applied to too few arguments, meaning that List.length arguments < operator_arity.

exception Leftover_expressions of operand Support.List2.t

Leftover_expressions (List2.T (o1, o2, os)) is raised from shunting_yard primitives if operands o2 :: os could not be reduced further, either because an operator is missing, or too many arguments were supplied to an operator.

val shunting_yard : primitive list -> operand

shunting_yard primitives is the result of rewriting primitives into a single operand, accounting for operator arity, precedence, fixity and associativity using Dijkstra's shunting yard algorithm.

raises Empty_expression

raises Misplaced_operator

raises Ambiguous_operator_placement

raises Consecutive_non_associative_operators

raises Arity_mismatch

raises Leftover_expressions