Module type Shunting_yard.OPERATOR

type associativity

type fixity

type operand

type t

The type of operator descriptors.

Destructors

val arity : t -> int

arity operator is the number of arguments expected for operator.

arity operator >= 0.

If operator is an infix operator, then arity operator = 2.

val precedence : t -> int

precedence operator is the order of operation value for operator.

If precedence o1 < precedence o2, then o2 precedes o1 such that a o1 b o2 c is rewritten to a o1 (b o2 c), where a, b and c are operands.

val fixity : t -> fixity

fixity operator is the fixity of operator.

val associativity : t -> associativity

associativity operator is the associativity of operator.

If the operator is prefix, then it is right-associative.

If the operator is postifx, then it is left-associative.

Writing

val write : t -> operand list -> operand

write operator arguments takes the list of arguments arguments and rewrites it to an application with head operator and arguments arguments.

The operands in arguments are in order, meaning that writing the prefix operator operator a1 a2 ... an has arguments = a1 :: a2 :: ... :: an.

Additionally, arity operator = List.length arguments.

Instances

Instance of equality such that equal a b = true if and only if a and b refer to the same operator. The arity, precedence, fixity and associativity of an operator is not sufficient to distinguish different operators.

This is used to determine whether non-associative operators appear consecutively during rewriting of an expression list.

include Support.Eq.EQ with type t := t

val equal : t -> t -> bool

equal a b is true if and only if a and b are equal. This should satisfy the following properties:

Reflexivity: equal a a = true

Symmetry: equal a b is equivalent to equal b a

Transitivity: if equal a b = true and equal b c = true, then equal a c = true

val (=) : t -> t -> bool

Operator alias of equal.

val (<>) : t -> t -> bool

Negation of operator (=).