Module Common.Error

val raise : exn -> 'a

exception Violation of string

val violation : string -> 'a

Raises a Violation exception with the given message.

exception NotImplemented of Location.t option * string

val not_implemented : Location.t -> string -> 'a

Raises a NotImplemented exception at the given location, with the given message.

val not_implemented' : string -> 'a

Raises a NotImplemented exception with the given message.

val located_exception : Location.t Support.List1.t -> exn -> exn

location_exception locations cause is a decorated exception having cause and locations for source file error-reporting.

val located_exception1 : Location.t -> exn -> exn

location_exception1 location cause is a decorated exception having cause and locations [location] for source file error-reporting.

val located_exception2 : Location.t -> Location.t -> exn -> exn

location_exception2 location1 location2 cause is a decorated exception having cause and locations [location1; location2] for source file error-reporting.

val raise_at : Location.t Support.List1.t -> exn -> 'a

raise_at locations cause raises the exception Located_exception { locations; cause }.

val raise_at1 : Location.t -> exn -> 'a

raise_at1 location cause raises the exception Located_exception { locations = [location]; cause }.

val raise_at2 : Location.t -> Location.t -> exn -> 'a

raise_at2 location1 location2 cause raises the exception Located_exception { locations = [location1; location2]; cause }.

val composite : exn Support.List2.t -> exn

composite causes is the composite error having many causes.

val composite2 : exn -> exn -> exn

composite2 cause1 cause2 is the composite error having causes [cause1; cause2].

type print_result

Abstract dummy datatype to enforce that printing be done using the printing functions provided by this module.

val register_printer : (exn -> print_result) -> unit

Wrapper around Printexc.register_printer. The given function should use partial pattern matching (ew) to signal that it doesn't know how to deal with other types of exceptions.

val register_printer' : (exn -> print_result option) -> unit

Wrapper around Printexc.register_printer. The given function should return None if it doesn't know how to handle a given exception.

val register_printing_function : (exn -> 'a option) -> (Support.Format.formatter -> 'a -> unit) -> unit

Registers a printer for an exception using the a given exception selection function and exception formatter.

val register_located_printing_function : (exn -> (Location.t * 'a) option) -> (Support.Format.formatter -> 'a -> unit) -> unit

Registers a printer for an exception that carries a location using the given exception selection function and exception formatter.

val print : (Support.Format.formatter -> unit) -> print_result

Use suplied formatter for printing errors.

val print_with_location : Location.t -> (Support.Format.formatter -> unit) -> print_result

Use supplied formatter for printing errors decorated with location information.

val report_mismatch : Support.Format.formatter -> string -> string -> (Support.Format.formatter -> 'a -> unit) -> 'a -> string -> (Support.Format.formatter -> 'b -> unit) -> 'b -> unit

Helper function to construct an error message reporting a mismatch between something that was expected and what was actually encountered. e.g. a type mismatch or a context clash.

example: report_mismatch ppf "Type mismatch." "Expected type" pp_ty1 ty1 "Inferred type" pp_ty2 ty2

val reset_information : unit -> unit

val get_information : unit -> string

val add_information : string -> unit