Module Common.Parenthesizer

Utilities for creating expression pretty-printers with minimal parentheses.

Handling Parentheses

Parentheses are re-introduced during pretty-printing using the precedence ordering specified in the parser. Operator associativities also need to be considered to avoid adding extraneous parentheses.

val parenthesize : (Support.Format.formatter -> 'a -> Stdlib.Unit.t) -> Support.Format.formatter -> 'a -> Stdlib.Unit.t

parenthesize pp is pp with the addition of parentheses.

type ('precedence, 'term) parenthesizing_formatter = ('term -> 'precedence) -> parent_precedence:'precedence -> (Support.Format.formatter -> 'term -> Stdlib.Unit.t) -> Support.Format.formatter -> 'term -> Stdlib.Unit.t

A parenthesizing formatter is a pretty-printer for at term that performs an additional check to determine whether to parenthesize the term or not. This check is performed by computing the precedence of the term and comparing it with the precedence of the parent node in the AST.

module Make_parenthesizer (Precedence : Support.Ord.ORD) : sig ... end

Functor for parenthesizing helper formatters.