Module Harpoon.CidProgRewrite

Contains a family a functions for rewriting references to program functions.

This is used after translation to replace all references for recursive calls with calls to the translated functions.

type cid_map = Beluga_syntax.Id.cid_prog -> Beluga_syntax.Id.cid_prog

val exp : cid_map -> Beluga_syntax.Synint.Comp.exp -> Beluga_syntax.Synint.Comp.exp