Module Harpoon.HarpoonState

type t

type triple = Session.t * Theorem.t * Beluga.Syntax.Int.Comp.proof_state

type triple_error = [

| `no_session

| `no_theorem of Session.t

| `no_subgoal of Session.t * Theorem.t

]

type triple_status = (triple_error, triple) Support.Either.t

val make : Options.save_mode -> Options.interaction_mode -> string -> string list -> IO.t -> Beluga.Syntax.Int.Comp.open_subgoal list -> t

val parsed_prompt : t -> ?source:string -> string -> string option -> 'a Beluga_parser.Simple.t -> 'a

Wraps IO.parsed_prompt.

val printf : t -> ('a, Support.Format.formatter, unit) Stdlib.format -> 'a

Prints a message to the user.

val session_list : t -> Session.t list

Gets a list of all sessions in the prover.

val defer_session : t -> unit

Moves the current session to the bottom of the session stack, selecting the next one.

val session_configuration_wizard : t -> bool

Configures a new session interactively and adds it to the state. Returns whether the new session was indeed added. A session is not added when the user aborts the configuration process.

val select_theorem : t -> Beluga.Name.t -> bool

val automation_state : t -> Automation.State.t

val interaction_mode : t -> Options.interaction_mode

Gets the interaction mode of the prover.

val next_triple : t -> triple_status

val completeness : t -> [ `complete | `incomplete ]

Decides whether there are any sessions in the theorem.

val on_session_completed : t -> Session.t -> unit

To be called by the REPL when a session is completed. This function analyzes the current save_mode and proceeds accordingly.

val serialize : t -> triple -> unit

Serializes the current state back to the loaded signature, preserving prover focus on the given state triple.

val fmt_ppr_session_list : Support.Format.formatter -> t -> unit

Prints a vertical, enumerated list of all sessions in this state, together with every theorem within each session.