Module Harpoon.IO

Centralizes high-level Harpoon input-output capabilities.

module Error : sig ... end

type t

Type of IO capabilities. Includes:

prompting the user for input,

displaying a message to the user.

When a function needs user interactivity, pass this object to it. If it only needs an output capability, see formatter.

val make : InputPrompt.t -> Stdlib.Format.formatter -> t

Constructs an IO capability.

val formatter : t -> Stdlib.Format.formatter

Extracts the formatter from an IO capability.

val prompt_number : t -> int

Gets the number of the last prompt that was shown.

val printf : t -> ('a, Stdlib.Format.formatter, unit) Stdlib.format -> 'a

Displays a formatted message.

val default_prompt_source : string

Default value used for the filename when displaying parse errors.

val parsed_prompt : ?source:string -> t -> string -> string option -> 'a Beluga_parser.Simple.t -> 'a

Prompts the user for input by displaying the given message. The response is parsed using the given parser. In case of parse errors, the prompt is repeated.

To allow the user to abort the loop, use a parser that accepts the empty string. See Parser.maybe.

The default value of the optional `source` parameter is default_source_prompt.