Module Harpoon.Options

type interaction_mode = [

| `stop

| `go_on

]

The `stop and `go_on flag control what happens in the presence of errors. In particular, the `stop flag will cause Harpoon to exit as soon as an error in encountered instead of continuing to process commands which may not make sense anymore. This is especially important when running tests.

type save_mode = [

| `save_back

| `no_save_back

]

Controls the behaviour of saving sessions back to the signature when they are completed.

type test_kind = [

| `incomplete

| `complete

]

type test_file = string * test_kind

type ('a, 'b) t = {

path : 'a;

all_paths : 'b;

test_file : test_file option;

test_start : int option;

test_stop : interaction_mode;

load_holes : bool;

save_back : save_mode;

}

type parsed_t

type elaborated_t = (string, string list) t

val parse_arguments : string list -> parsed_t

val elaborate : parsed_t -> elaborated_t