Module Harpoon.Revisit

val revisit_disambiguation : Beluga_syntax.Synint.Sgn.sgn -> Beluga_parser.Disambiguation_state.Disambiguation_state.state Beluga_syntax.Id.Prog.Hashtbl.t * Beluga_parser.Disambiguation_state.Disambiguation_state.state

revisit_disambiguation sgn is a lookup table for the disambiguation state at each Harpoon proof in sgn, and a snapshot of the disambiguation state at the end of sgn. These disambiguation states are snapshots, and should be used with a checkpoint to disambiguate new parsemes.

val revisit_indexing : Beluga_syntax.Synint.Sgn.sgn -> Beluga.Index_state.Indexing_state.state Beluga_syntax.Id.Prog.Hashtbl.t * Beluga.Index_state.Indexing_state.state

revisit_indexing sgn is a lookup table for the indexing state at each Harpoon proof in sgn, and a snapshot of the indexing state at the end of sgn. These indexing states are snapshots, and should be used with a checkpoint to index new ASTs.