Module Harpoon.Theorem

module CompS = Beluga.Store.Cid.Comp

module Conf : sig ... end

module Action : sig ... end

module Direction : sig ... end

type t

type theorem = t

type 'a subgoal_hook = Beluga.Syntax.Int.Comp.proof_state -> 'a

val printf : t -> ('a, Stdlib.Format.formatter, unit) Stdlib.format -> 'a

TODO hide this from outside of this module

val get_entry' : t -> Beluga.Id.cid_prog * CompS.Entry.t

Gets the cid and Store entry for this theorem.

val get_history_names : t -> string list * string list

get_history_names t = (as1, as2) where as1 are the past history items and as2 are the future history items.

val get_cid : t -> Beluga.Id.cid_prog

Gets the cid for this theorem.

val get_entry : t -> CompS.Entry.t

Gets the Store entry for this theorem.

val get_name : t -> Beluga.Name.t

val has_name_of : t -> Beluga.Name.t -> bool

Checks if the theorem's name is equal to the given name.

val has_cid_of : t -> Beluga.Id.cid_prog -> bool

Checks if the theorem's cid is equal to the given cid.

val get_statement : t -> Beluga.Syntax.Int.Comp.tclo

Gets the type of the initial subgoal of this theorem. Note that this is the statement of the theorem _including_ inductive stars. To get the statement without stars, retrieve the type of the theorem in the store via its cid.

val serialize : Stdlib.Format.formatter -> t -> unit

val next_subgoal : t -> Beluga.Syntax.Int.Comp.proof_state option

val select_subgoal_satisfying : t -> (Beluga.Syntax.Int.Comp.proof_state -> bool) -> Beluga.Syntax.Int.Comp.proof_state option

val dump_proof : Stdlib.Format.formatter -> t -> unit

val show_proof : t -> unit

val show_subgoals : t -> unit

val defer_subgoal : t -> unit

val subgoals : t -> Beluga.Syntax.Int.Comp.proof_state list

val count_subgoals : t -> int

val apply : t -> Action.t -> unit

Executes the given action on this theorem, and records the action to the theorem's history.

val history_step : t -> Direction.t -> bool

Inverts the last action recorded in the history. Returns false if the history is empty and no action can be inverted.

val apply_subgoal_replacement : t -> string -> Beluga.Syntax.Int.Comp.proof_state -> (Beluga.Syntax.Int.Comp.proof -> Beluga.Syntax.Int.Comp.proof) -> Beluga.Syntax.Int.Comp.proof_state -> unit

Replaces the subgoal with another, solving it by transforming an incomplete proof for the new subgoal. Interally this uses `apply` and so records theorem history, using the given action name.

val rename_variable : Beluga.Name.t -> Beluga.Name.t -> [ `comp | `meta ] -> t -> Beluga.Syntax.Int.Comp.proof_state -> bool

Renames the given variable at the given level. Returns true iff such a variable could be renamed. Else, there was no such variable.

val configure : Beluga.Id.cid_comp_const -> Stdlib.Format.formatter -> (t -> unit subgoal_hook) list -> Beluga.Syntax.Int.Comp.proof_state -> Beluga.Syntax.Int.Comp.proof_state list -> t

val configure_set : Stdlib.Format.formatter -> (t -> unit subgoal_hook) list -> Conf.t list -> Beluga.Id.cid_mutual_group * t list

type completeness = [

| `incomplete

| `complete

]

val completeness : t -> completeness

Decides whether the theorem is complete.

val materialize : t -> unit

Assigns a declaration number to this theorem.

Be careful calling this. It should never be the case that some theorems in a session are materialized and some not.

This function is idempotent: calling this on a theorem that is already materialized will do nothing.