Module Harpoon.Translate

type error =

| IncompleteProof

exception Error of error

type result = (error, Synint.Comp.exp) Support.Either.t

val entry : Beluga.Store.Cid.Comp.Entry.t -> result

Translates a theorem given by a Store entry.

val theorem : Synint.Comp.thm -> Synint.Comp.typ -> Synint.Comp.exp

Translates a theorem. Theorems proven already by a program are returned as-is, but theorems proven with a Harpoon proof are translated.

val trap : (unit -> 'a) -> (error, 'a) Support.Either.t

Traps exceptions raised by this module.

val fmt_ppr_error : Support.Format.formatter -> error -> unit

Formats an error.

val fmt_ppr_result : Support.Format.formatter -> result -> unit

Formats a translation result.