Module Support

Utility libraries.

Monads and Modular Type Classes

module Monad : sig ... end

Definition of monadic types, which are abstract descriptors of computations.

module Functor : sig ... end

module Apply : sig ... end

module Alternative : sig ... end

module State : sig ... end

State monad definition.

module Show : sig ... end

module Eq : sig ... end

Types having an equality operator.

module Ord : sig ... end

Totally ordered types.

module Hash : sig ... end

Hashable types.

module Imperative_state : sig ... end

Mutable state helper definitions.

module Format_state : sig ... end

Format combinators using the Format module, but using a state having a Format.formatter state.

Extensions to Imported Libraries

module DynArray : sig ... end

module Format : sig ... end

module Gen : sig ... end

module Hashtbl : sig ... end

module Int : sig ... end

module List : sig ... end

module Option : sig ... end

module Seq : sig ... end

module Stack : sig ... end

module String : sig ... end

Additional Data Structures

module PureStack : sig ... end

Persistent stack data structure.

module Either : sig ... end

Ad hoc disjoint unions of two types.

module Pair : sig ... end

module List1 : sig ... end

module List2 : sig ... end

module History : sig ... end

Utilities

module Equality : sig ... end

module Misc : sig ... end

Totally miscellaneous functions.

module Debug : sig ... end

Debug-print library with up to 31 separate categories of information.

module Fun : sig ... end

module Files : sig ... end

File-handling utilities