Module Support.Files

Miscellaneous file handling utilities

val read_lines : string -> string list

read_lines filename is the list of '\n' separated lines read from the file at filename.

val with_temp_file : string -> string -> (string -> Stdlib.out_channel -> 'a) -> 'a

with_temp_file temp_dir basename f opens a temporary file basename in directory temp_dir and runs f with the path and output channel to that file. This function guarantees that even in case of exception, the output channel is closed and the temporary file is unlinked.

val with_in_channel : string -> (Stdlib.in_channel -> 'a) -> 'a

with_in_channel filename f opens an input channel on filename and runs f with it. Guarantees to close the channel, even in case of exceptional exit.