Parameter Make.Key

type t

The type of totally ordered and hashable types.

Instances

include Eq.EQ with type t := t

val equal : t -> t -> bool

equal a b is true if and only if a and b are equal. This should satisfy the following properties:

Reflexivity: equal a a = true

Symmetry: equal a b is equivalent to equal b a

Transitivity: if equal a b = true and equal b c = true, then equal a c = true

val (=) : t -> t -> bool

Operator alias of equal.

val (<>) : t -> t -> bool

Negation of operator (=).

include Hash.HASH with type t := t

val hash : t -> int

hash e is the hash code computed for e.

If e1 and e2 as equal, then hash e1 and hash e2 must be equal.