Module type Hash.HASH

Module type for hashable types.

Hashable types may be used as keys for Hashtbl.Make.

type t

val hash : t -> int

hash e is the hash code computed for e.

If e1 and e2 as equal, then hash e1 and hash e2 must be equal.