Module Support.Imperative_state

Mutable state helper definitions.

This is analogous to the state monad defined in State, but the imperative state type assumes that the state is mutable.

module type IMPERATIVE_STATE = sig ... end

Abstract definition of helpers for stateful operations using a strictly mutable state.

module Make (State : sig ... end) : IMPERATIVE_STATE with type state = State.state

Functor building an implementation of IMPERATIVE_STATE for a given type of states.