Module type Imperative_state.IMPERATIVE_STATE

Abstract definition of helpers for stateful operations using a strictly mutable state.

type state

val traverse_tuple2 : state -> (state -> 'a1 -> 'b1) -> (state -> 'a2 -> 'b2) -> ('a1 * 'a2) -> 'b1 * 'b2

val traverse_tuple3 : state -> (state -> 'a1 -> 'b1) -> (state -> 'a2 -> 'b2) -> (state -> 'a3 -> 'b3) -> ('a1 * 'a2 * 'a3) -> 'b1 * 'b2 * 'b3

val traverse_list : state -> (state -> 'a -> 'b) -> 'a List.t -> 'b List.t

traverse_list state f [x1; x2; ...; xn] is [y1; y2; ...; yn] where yi = f state xi, and y1, y2, ..., yn are computed in order, meaning that y1 is computed first, then y2, etc.

val traverse_list1 : state -> (state -> 'a -> 'b) -> 'a List1.t -> 'b List1.t

val traverse_list2 : state -> (state -> 'a -> 'b) -> 'a List2.t -> 'b List2.t

val iter_list : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a List.t -> Stdlib.Unit.t

iter_list state f [x1; x2; ...; xn] is f state x1; f state x2; ...; f state xn.

val iter_list1 : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a List1.t -> Stdlib.Unit.t

val iter_list2 : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a List2.t -> Stdlib.Unit.t

val traverse_reverse_list : state -> (state -> 'a -> 'b) -> 'a List.t -> 'b List.t

traverse_reverse_list state f [x1; x2; ...; xn] is [y1; y2; ...; yn] where yi = f state xi, and y1, y2, ..., yn are computed in reverse order, meaning that yn is computed first, then y(n-1), etc.

val traverse_reverse_list1 : state -> (state -> 'a -> 'b) -> 'a List1.t -> 'b List1.t

val traverse_reverse_list2 : state -> (state -> 'a -> 'b) -> 'a List2.t -> 'b List2.t

val iter_rev_list : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a List.t -> Stdlib.Unit.t

iter_rev_list state f [x1; x2; ...; xn] is f state xn; f state x(n-1); ...; f state x1.

val iter_rev_list1 : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a List1.t -> Stdlib.Unit.t

val iter_rev_list2 : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a List2.t -> Stdlib.Unit.t

val traverse_option : state -> (state -> 'a -> 'b) -> 'a Option.t -> 'b Option.t

val iter_option : state -> (state -> 'a -> Stdlib.Unit.t) -> 'a Option.t -> Stdlib.Unit.t

val seq_list : state -> (state -> 'a) List.t -> 'a List.t

val seq_list1 : state -> (state -> 'a) List1.t -> 'a List1.t

val iter_seq : state -> (state -> Stdlib.Unit.t) List.t -> Stdlib.Unit.t