Module Support.Misc

Totally miscellaneous functions.

include module type of struct include Equality end

val (=) : int -> int -> bool

A restricted form of the equality operator that works only for integers.

Fully polymorphic equality is dangerous and limits refactorability of the code, so every module should open this one to avoid using polymorphic equality.

val noexcept : (unit -> unit) -> unit

Runs a function ignoring all exceptions. In general this is a terrible idea, but it is sometimes necessary when performing cleanup that may fail while in an exception handler.

exception NotImplemented of string

An exception to be raised in unimplemented features. * Code that raises this exception should never be committed.

val not_implemented : string -> 'a

val enumerate_with_state : 's -> ('s0 -> 's0 * 'i) -> 'a list -> 's1 * ('i0 * 'a0) list

Enumerates a list using a state transformer to generate indices. The initial seed s contains the initial state and the function f transforms this state to compute a new state and an index. These indices are sequentially zipped with the given list to produce an indexed list, and the final state is returned.

val enumerate : 'a list -> (int * 'a0) list

Enumerates a list by pairing each element with its index.

val tuple : 'a -> 'b -> 'a0 * 'b0

Forms the tuple of its two inputs.

val throw : exn -> 'b -> 'a

Creates a constant function that raises the given exception. Useful when eliminating option-types.

val on : ('a -> 'b) -> ('b0 -> 'b0 -> 'c) -> 'a0 -> 'a0 -> 'c

type void = {

impossible : 'a. 'a;

}