Module Support.Monad

Definition of monadic types, which are abstract descriptors of computations.

For better performance, avoid using monads.

module type MONAD = sig ... end

The abstract datatype of actions.

module Make (Monad : sig ... end) : MONAD with type 'a t = 'a Monad.t

Functor building the aliases for a minimal implementation for MONAD.