Module Pair.MakeEq

Parameters

module E1 : Eq.EQ

module E2 : Eq.EQ

Signature

type t = (E1.t, E2.t) t

The type of elements to check for equality.

val equal : t -> t -> bool

equal a b is true if and only if a and b are equal. This should satisfy the following properties:

Reflexivity: equal a a = true

Symmetry: equal a b is equivalent to equal b a

Transitivity: if equal a b = true and equal b c = true, then equal a c = true

val (=) : t -> t -> bool

Operator alias of equal.

val (<>) : t -> t -> bool

Negation of operator (=).