Module Support.Show

module type SHOW = sig ... end

Module type for types with functions for pretty-printing and converting values to readable strings.

module Make (T : sig ... end) : SHOW with type t = T.t

Functor building an implementation of SHOW given a type with a pretty-printer.

val make : (Format.formatter -> 't -> unit) -> (module SHOW with type t = 't)

make pp is equivalent to the functor Make (T) with T.pp = pp.

val contramap : (module SHOW with type t = 's) -> ('t -> 's) -> (module SHOW with type t = 't)

If val f : 't -> 's, then contramap show f is an instance of SHOW for values of type 't by the SHOW insance show for values of type 's.