Module Syncom

Self-contained syntactic elements common to one or more of Beluga's syntaxes.

Operators

module Associativity : sig ... end

Description for the associativity of user-defined operators.

module Fixity : sig ... end

Description for the fixity of user-defined operators.

module Operator : sig ... end

Description of user-defined operators.

Annotations

module Depend : sig ... end

Dependency annotations to parameters to dependent type functions.

module Inductivity : sig ... end

Annotations for parameters and variables generating induction hypotheses.

module Plicity : sig ... end

Annotations for implicit or explicit parameters.

Pretty-Printing

module Parenthesizer : sig ... end

Utilities for creating expression pretty-printers with minimal parentheses.

Identifiers

module Identifier : sig ... end

String identifiers.

module Qualified_identifier : sig ... end

Namespaced identifiers.

module Binding_tree : sig ... end

Mutable map data structure from namespaced identifiers to values.

module Id : sig ... end

Constant IDs for referring to constants in the global mutable store.

module HoleId : sig ... end

Legacy Names

module Name : sig ... end

module Gensym : sig ... end

Utilities

module Error : sig ... end

Utilities for raising and catching exceptions.

module Position : sig ... end

Character positions in source files.

module Location : sig ... end

Character position ranges in source files.