Module Syncom.HoleId

type name =

| Anonymous

| Named of string

type t

val of_int : int -> t

val string_of_hole_id : t -> string

module Make () : sig ... end

Construct a counter for hole id numbers.

val name_of_option : string option -> name

Converts an option type to a name.

val option_of_name : name -> string option

Converts a name to an option type.

val string_of_name : name -> string

Stringifies a hole name.

If the hole is anonymous, then its name is the empty string.

val string_of_name_or_id : (name * t) -> string

Stringifies a hole name.

If the hole is anonymous, then use its hole id.

val is_anonymous : name -> bool

Decides whether a hole name is anonymous.

val fmt_ppr_name : Stdlib.Format.formatter -> name -> unit

val fmt_ppr_id : Stdlib.Format.formatter -> t -> unit

val compare : t -> t -> int