Module Synext

External Syntax Definition

External LF Syntax

module LF : sig ... end

The representation of LF kinds, types, and terms after parsing and data-dependent disambiguation.

External Contextual LF Syntax

module CLF : sig ... end

The representation of contextual LF types, terms, and patterns after parsing and data-dependent disambiguation.

External Meta-Syntax

module Meta : sig ... end

The representation of meta-types, meta-objects, meta-object patterns, meta-substitutions and meta-contexts after parsing and data-dependent disambiguation.

External Computational Syntax

module Comp : sig ... end

The representation of computation-level kinds, types, expressions and patterns after parsing and data-dependent disambiguation.

External Harpoon Syntax

module Harpoon : sig ... end

External Signature Syntax

module Signature : sig ... end

Type Aliases

LF

type lf_kind = LF.Kind.t

type lf_typ = LF.Typ.t

type lf_term = LF.Term.t

Contextual LF

type clf_typ = CLF.Typ.t

type clf_term = CLF.Term.t

type clf_term_pattern = CLF.Term.Pattern.t

type clf_substitution = CLF.Substitution.t

type clf_substitution_pattern = CLF.Substitution.Pattern.t

type clf_context = CLF.Context.t

type clf_context_pattern = CLF.Context.Pattern.t

Meta-level

type meta_typ = Meta.Typ.t

type meta_object = Meta.Object.t

type meta_pattern = Meta.Pattern.t

type schema = Meta.Schema.t

type meta_context = Meta.Context.t

Computation-level

type comp_kind = Comp.Kind.t

type comp_typ = Comp.Typ.t

type comp_expression = Comp.Expression.t

type comp_pattern = Comp.Pattern.t

type comp_copattern = Comp.Copattern.t

type comp_context = Comp.Context.t

Harpoon

type harpoon_command = Harpoon.Command.t

type harpoon_directive = Harpoon.Directive.t

type harpoon_proof = Harpoon.Proof.t

type harpoon_hypothetical = Harpoon.Hypothetical.t

type harpoon_repl_command = Harpoon.Repl.Command.t

type harpoon_split_branch = Harpoon.Split_branch.t

type harpoon_split_branch_label = Harpoon.Split_branch.Label.t

type harpoon_suffices_branch = Harpoon.Suffices_branch.t

Signature

type signature_pragma = Signature.Pragma.t

type signature_global_pragma = Signature.Global_pragma.t

type signature_totality_declaration = Signature.Totality.Declaration.t

type 'argument signature_totality_order = 'argument Signature.Totality.Order.t

type signature_declaration = Signature.Declaration.t

type signature_entry = Signature.Entry.t

type signature = Signature.t

Common

include module type of struct include Common end

Operators

module Associativity = Common.Associativity

module Fixity = Common.Fixity

module Operator = Common.Operator

Parameter Annotations

module Plicity = Common.Plicity

module Inductivity = Common.Inductivity

Pretty-Printing

module Parenthesizer = Common.Parenthesizer

Identifiers

module Identifier = Common.Identifier

module Qualified_identifier = Common.Qualified_identifier

Utilities

module Error = Common.Error

module Position = Common.Position

module Location = Common.Location

Locations

LF

val location_of_lf_kind : Synext__.Synext_definition.LF.Kind.t -> Common.Location.t

val location_of_lf_typ : Synext__.Synext_definition.LF.Typ.t -> Common.Location.t

val location_of_lf_term : Synext__.Synext_definition.LF.Term.t -> Common.Location.t

Contextual LF

val location_of_clf_typ : Synext__.Synext_definition.CLF.Typ.t -> Common.Location.t

val location_of_clf_term : Synext__.Synext_definition.CLF.Term.t -> Common.Location.t

val location_of_clf_substitution : Synext__.Synext_definition.CLF.Substitution.t -> Common.Location.t

val location_of_clf_substitution_pattern : Synext__.Synext_definition.CLF.Substitution.Pattern.t -> Common.Location.t

val location_of_clf_context : Synext__.Synext_definition.CLF.Context.t -> Common.Location.t

val location_of_clf_context_pattern : Synext__.Synext_definition.CLF.Context.Pattern.t -> Common.Location.t

val location_of_clf_term_pattern : Synext__.Synext_definition.CLF.Term.Pattern.t -> Common.Location.t

Meta-Level

val location_of_meta_type : Synext__.Synext_definition.Meta.Typ.t -> Common.Location.t

val location_of_meta_object : Synext__.Synext_definition.Meta.Object.t -> Common.Location.t

val location_of_meta_object_pattern : Synext__.Synext_definition.Meta.Pattern.t -> Common.Location.t

val location_of_meta_context : Synext__.Synext_definition.Meta.Context.t -> Common.Location.t

val location_of_schema : Synext__.Synext_definition.Meta.Schema.t -> Common.Location.t

Computation-Level

val location_of_comp_kind : Synext__.Synext_definition.Comp.Kind.t -> Common.Location.t

val location_of_comp_typ : Synext__.Synext_definition.Comp.Typ.t -> Common.Location.t

val location_of_comp_expression : Synext__.Synext_definition.Comp.Expression.t -> Common.Location.t

val location_of_comp_pattern : Synext__.Synext_definition.Comp.Pattern.t -> Common.Location.t

val location_of_comp_copattern : Synext__.Synext_definition.Comp.Copattern.t -> Common.Location.t

Harpoon

val location_of_harpoon_proof : Synext__.Synext_definition.Harpoon.Proof.t -> Common.Location.t

val location_of_harpoon_command : Synext__.Synext_definition.Harpoon.Command.t -> Common.Location.t

val location_of_harpoon_directive : Synext__.Synext_definition.Harpoon.Directive.t -> Common.Location.t

val location_of_harpoon_split_branch : Synext__.Synext_definition.Harpoon.Split_branch.t -> Common.Location.t

val location_of_harpoon_split_branch_label : Synext__.Synext_definition.Harpoon.Split_branch.Label.t -> Common.Location.t

val location_of_harpoon_suffices_branch : Synext__.Synext_definition.Harpoon.Suffices_branch.t -> Common.Location.t

val location_of_harpoon_hypothetical : Synext__.Synext_definition.Harpoon.Hypothetical.t -> Common.Location.t

val location_of_harpoon_repl_command : Synext__.Synext_definition.Harpoon.Repl.Command.t -> Common.Location.t

Signature

val location_of_signature_pragma : Synext__.Synext_definition.Signature.Pragma.t -> Common.Location.t

val location_of_signature_global_pragma : Synext__.Synext_definition.Signature.Global_pragma.t -> Common.Location.t

val location_of_signature_totality_declaration : Synext__.Synext_definition.Signature.Totality.Declaration.t -> Common.Location.t

val location_of_signature_totality_order : 'a Synext__.Synext_definition.Signature.Totality.Order.t -> Common.Location.t

val location_of_signature_declaration : Synext__.Synext_definition.Signature.Declaration.t -> Common.Location.t

Precedences

Precedence of LF Syntax

module Lf_precedence : sig ... end

val precedence_of_lf_kind : Synext__.Synext_definition.LF.Kind.t -> Lf_precedence.t

val precedence_of_lf_typ : Synext__.Synext_definition.LF.Typ.t -> Lf_precedence.t

val precedence_of_lf_term : Synext__.Synext_definition.LF.Term.t -> Lf_precedence.t

Precedence of Contextual LF Syntax

module Clf_precedence : sig ... end

val precedence_of_clf_typ : Synext__.Synext_definition.CLF.Typ.t -> Clf_precedence.t

val precedence_of_clf_term : Synext__.Synext_definition.CLF.Term.t -> Clf_precedence.t

val precedence_of_clf_term_pattern : Synext__.Synext_definition.CLF.Term.Pattern.t -> Clf_precedence.t

Precedence of Meta-Level Syntax

module Meta_precedence : sig ... end

val precedence_of_schema : Synext__.Synext_definition.Meta.Schema.t -> Meta_precedence.t

module Comp_precedence : sig ... end

Precedence of Computation-Level Syntax

val precedence_of_comp_kind : Synext__.Synext_definition.Comp.Kind.t -> Comp_precedence.t

val precedence_of_comp_typ : Synext__.Synext_definition.Comp.Typ.t -> Comp_precedence.t

val precedence_of_comp_expression : Synext__.Synext_definition.Comp.Expression.t -> Comp_precedence.t

val precedence_of_comp_pattern : Synext__.Synext_definition.Comp.Pattern.t -> Comp_precedence.t

val precedence_of_comp_copattern : Synext__.Synext_definition.Comp.Copattern.t -> Comp_precedence.t

Pretty-Printing

LF

val pp_lf_kind : Stdlib.Format.formatter -> Synext__.Synext_definition.LF.Kind.t -> unit

val pp_lf_typ : Stdlib.Format.formatter -> Synext__.Synext_definition.LF.Typ.t -> unit

val pp_lf_term : Stdlib.Format.formatter -> Synext__.Synext_definition.LF.Term.t -> unit

Contextual LF

val pp_clf_typ : Stdlib.Format.formatter -> Synext__.Synext_definition.CLF.Typ.t -> unit

val pp_clf_term : Stdlib.Format.formatter -> Synext__.Synext_definition.CLF.Term.t -> unit

val pp_clf_term_pattern : Stdlib.Format.formatter -> Synext__.Synext_definition.CLF.Term.Pattern.t -> unit

val pp_clf_substitution : Stdlib.Format.formatter -> Synext__.Synext_definition.CLF.Substitution.t -> unit

val pp_clf_substitution_pattern : Stdlib.Format.formatter -> Synext__.Synext_definition.CLF.Substitution.Pattern.t -> unit

val pp_clf_context : Stdlib.Format.formatter -> Synext__.Synext_definition.CLF.Context.t -> unit

val pp_clf_context_pattern : Stdlib.Format.formatter -> Synext__.Synext_definition.CLF.Context.Pattern.t -> unit

Meta-Level

val pp_meta_typ : Stdlib.Format.formatter -> Synext__.Synext_definition.Meta.Typ.t -> unit

val pp_meta_object : Stdlib.Format.formatter -> Synext__.Synext_definition.Meta.Object.t -> unit

val pp_meta_context : Stdlib.Format.formatter -> Synext__.Synext_definition.Meta.Context.t -> unit

val pp_meta_pattern : Stdlib.Format.formatter -> Synext__.Synext_definition.Meta.Pattern.t -> unit

val pp_schema : Stdlib.Format.formatter -> Synext__.Synext_definition.Meta.Schema.t -> unit

Computation-Level

val pp_comp_kind : Stdlib.Format.formatter -> Synext__.Synext_definition.Comp.Kind.t -> unit

val pp_comp_typ : Stdlib.Format.formatter -> Synext__.Synext_definition.Comp.Typ.t -> unit

val pp_comp_expression : Stdlib.Format.formatter -> Synext__.Synext_definition.Comp.Expression.t -> unit

val pp_comp_pattern : Stdlib.Format.formatter -> Synext__.Synext_definition.Comp.Pattern.t -> unit

val pp_comp_copattern : Stdlib.Format.formatter -> Synext__.Synext_definition.Comp.Copattern.t -> unit

val pp_comp_context : Stdlib.Format.formatter -> Synext__.Synext_definition.Comp.Context.t -> unit

Harpoon

val pp_harpoon_proof : Stdlib.Format.formatter -> Synext__.Synext_definition.Harpoon.Proof.t -> unit

val pp_harpoon_command : Stdlib.Format.formatter -> Synext__.Synext_definition.Harpoon.Command.t -> unit

val pp_harpoon_directive : Stdlib.Format.formatter -> Synext__.Synext_definition.Harpoon.Directive.t -> unit

val pp_harpoon_hypothetical : Stdlib.Format.formatter -> Synext__.Synext_definition.Harpoon.Hypothetical.t -> unit

val pp_harpoon_repl_command : Stdlib.Format.formatter -> Synext__.Synext_definition.Harpoon.Repl.Command.t -> unit

Signature

val pp_signature_pragma : Stdlib.Format.formatter -> Synext__.Synext_definition.Signature.Pragma.t -> unit

val pp_signature_global_pragma : Stdlib.Format.formatter -> Synext__.Synext_definition.Signature.Global_pragma.t -> unit

val pp_signature_totality_declaration : Stdlib.Format.formatter -> Synext__.Synext_definition.Signature.Totality.Declaration.t -> unit

val pp_signature_declaration : Stdlib.Format.formatter -> Synext__.Synext_definition.Signature.Declaration.t -> unit

val pp_signature_entry : Stdlib.Format.formatter -> Synext__.Synext_definition.Signature.Entry.t -> unit

val pp_signature : Stdlib.Format.formatter -> Synext__.Synext_definition.Signature.t -> unit