Module Synext.CLF

The representation of contextual LF types, terms, and patterns after parsing and data-dependent disambiguation.

ASTs constructed with the constructors in this module are not necessarily in normal form.

The distinction between contextual LF objects and plain LF objects is that contextual LF objects may have substitutions, and may appear in patterns. Plain LF objects are only used in the definition of type-level or term-level LF constants.

These types are suited for pretty-printing and elaboration to the internal syntax. Note that this is a named representation.

The metavariable:

x ranges over variables

c ranges over term-level constants

a ranges over type-level constants

s ranges over substitutions

g ranges over contexts

id ranges over identifiers

n ranges over integers

Contextual LF types           A, B ::= a | Πx:A.B | A → B | A M1 M2 ... Mn
                                         | block (x1:A1, x2:A2, ..., xn:An)
Contextual LF terms           M, N ::= c | x | #x | $x | λx:A.M | M N1 N2 ... Nn
                                         | M:A | M[σ]
                                         | _ | ? | ?id | <M1; M2; ...; Mn> | M.n | M.id
Contextual LF substitutions   σ    ::= ^ | … | σ, M | s[σ]
Contextual LF contexts        Ψ    ::= ^ | g | Ψ, x:A

Contextual LF patterns                Mp, Np ::= c | x | #x | $x | λx:A.Mp
                                              | Mp Np1 Np2 ... Npn
                                              | Mp:A | Mp[σ] | _
                                              | <Mp1; Mp2; ...; Mpn>
                                              | Mp.n | Mp.id
Contextual LF substitution patterns   σp     ::= ^ | … | σp, Mp | s[σ]
Contextual LF context patterns        Ψp     ::= ^ | g | Ψp, x:A

The representation of contextual LF types, terms, and patterns after parsing and data-dependent disambiguation.

ASTs constructed with the constructors in this module are not necessarily in normal form.

The distinction between contextual LF objects and plain LF objects is that contextual LF objects may have substitutions, and may appear in patterns. Plain LF objects are only used in the definition of type-level or term-level LF constants.

These types are suited for pretty-printing and elaboration to the internal syntax. Note that this is a named representation.

The metavariable:

x ranges over variables

c ranges over term-level constants

a ranges over type-level constants

s ranges over substitutions

g ranges over contexts

id ranges over identifiers

n ranges over integers

Contextual LF types           A, B ::= a | Πx:A.B | A → B | A M1 M2 ... Mn
                                         | block (x1:A1, x2:A2, ..., xn:An)
Contextual LF terms           M, N ::= c | x | #x | $x | λx:A.M | M N1 N2 ... Nn
                                         | M:A | M[σ]
                                         | _ | ? | ?id | <M1; M2; ...; Mn> | M.n | M.id
Contextual LF substitutions   σ    ::= ^ | … | σ, M | s[σ]
Contextual LF contexts        Ψ    ::= ^ | g | Ψ, x:A

Contextual LF patterns                Mp, Np ::= c | x | #x | $x | λx:A.Mp
                                              | Mp Np1 Np2 ... Npn
                                              | Mp:A | Mp[σ] | _
                                              | <Mp1; Mp2; ...; Mpn>
                                              | Mp.n | Mp.id
Contextual LF substitution patterns   σp     ::= ^ | … | σp, Mp | s[σ]
Contextual LF context patterns        Ψp     ::= ^ | g | Ψp, x:A

module Typ : sig ... end

External contextual LF types.

module Term : sig ... end

External contextual LF terms.

module Substitution : sig ... end

External contextual LF substitutions.

module Context : sig ... end

External contextual LF contexts.