Module Term.Pattern

External contextual LF term patterns.

type t =

| Variable of {

location : Common.Location.t;

identifier : Common.Identifier.t;

}
(*
Variable { identifier = "x"; _ } is the term-level variable pattern "x".
*)

| Parameter_variable of {

location : Common.Location.t;

identifier : Common.Identifier.t;

}
(*
Parameter_variable { identifier = "#x"; _ } is the term-level parameter variable pattern with name "#x".
*)

| Substitution_variable of {

location : Common.Location.t;

identifier : Common.Identifier.t;

}
(*
Substitution_variable { identifier = "$x"; _ } is the term-level substitution variable pattern with name "$x".
*)

| Constant of {

location : Common.Location.t;

identifier : Common.Qualified_identifier.t;

operator : Common.Operator.t;

quoted : Stdlib.Bool.t;

}
(*
Constant { identifier = "c"; _ } is the term-level constant pattern "c".
*)

| Wildcard of {

location : Common.Location.t;

}
(*
Wildcard _ is the term-level catch-all pattern _.
*)

| Tuple of {

location : Common.Location.t;

terms : t Support.List1.t;

}
(*
Tuple { terms; _ } is the tuple pattern <p1; p2; ...; pn> if terms = [p1; p2; ...; pn].
*)

| Projection of {

location : Common.Location.t;

term : t;

projection : [ `By_identifier of Common.Identifier.t | `By_position of Support.Int.t ];

}
(*
Projection { term; _ } is the pattern for the projection of a tuple term. This projection is used to indicate which element of a tuple the scrutinee is matched against in a case-expression.
*)

| Abstraction of {

location : Common.Location.t;

parameter_identifier : Common.Identifier.t Support.Option.t;

parameter_type : Typ.t Support.Option.t;

body : t;

}
(*
Abstraction { parameter_identifier = Option.Some "x"; parameter_type = Option.Some t; body; _ } is the pattern \x:t. body.
*)

| Substitution of {

location : Common.Location.t;

term : t;

substitution : Substitution.t;

}
(*
Substitution { term; substitution; _ } is the pattern term[substitution].
*)

| Application of {

location : Common.Location.t;

applicand : t;

arguments : t Support.List1.t;

}
(*
Application { applicand; arguments; _ } is the term-level application pattern of applicand with arguments.

If applicand = Term.Constant { operator; _ } and Operator.is_infix operator, then List.length arguments = 2.

If applicand = Term.Constant { operator; _ } and Operator.is_postfix operator, then List.length arguments = 1.

*)

| TypeAnnotated of {

location : Common.Location.t;

term : t;

typ : Typ.t;

}
(*
TypeAnnotated { term = x; typ = t; _ } is the pattern x : t.
*)