Module Synext.Comp

The representation of computation-level kinds, types, expressions and patterns after parsing and data-dependent disambiguation.

ASTs constructed with the constructors in this module are not necessarily in normal form.

These types are suited for pretty-printing and elaboration to the internal syntax. Note that this is a named representation.

The metavariable:

x ranges over computation-level variables

c ranges over computation-level constants

X ranges over meta-level variables

Computational kinds         K    ::= ctype | ΠX:T.K | T → K |
Computational types         T, S ::= ΠX:T.S | T → S | T × S
                                            | U | T C1 C2 ... CN
Computational expressions   E    ::= x | c | let x = P in E | impossible E
                                       | (E1, E2, ..., En) | ? | ?id | _
                                       | E1 E2 ... En
                                       | (case E of P1 => E1 | P2 => E2 | ... | Pn => En)
                                       | fn x1, x2, ..., xn => E
                                       | mlam X1, X2, ..., Xn => E
                                       | (fun P1 => E1 | P2 => E2 | ... | Pn => En)
Computational patterns      P    ::= x | c | Cp | (P1, P2, ..., Pn) | P1 P2 ... Pn
                                       | P : T | { X : U } P | _
Computational context       Γ    ::= ^ | Γ, x : T

The representation of computation-level kinds, types, expressions and patterns after parsing and data-dependent disambiguation.

ASTs constructed with the constructors in this module are not necessarily in normal form.

These types are suited for pretty-printing and elaboration to the internal syntax. Note that this is a named representation.

The metavariable:

x ranges over computation-level variables

c ranges over computation-level constants

X ranges over meta-level variables

Computational kinds         K    ::= ctype | ΠX:T.K | T → K |
Computational types         T, S ::= ΠX:T.S | T → S | T × S
                                            | U | T C1 C2 ... CN
Computational expressions   E    ::= x | c | let x = P in E | impossible E
                                       | (E1, E2, ..., En) | ? | ?id | _
                                       | E1 E2 ... En
                                       | (case E of P1 => E1 | P2 => E2 | ... | Pn => En)
                                       | fn x1, x2, ..., xn => E
                                       | mlam X1, X2, ..., Xn => E
                                       | (fun P1 => E1 | P2 => E2 | ... | Pn => En)
Computational patterns      P    ::= x | c | Cp | (P1, P2, ..., Pn) | P1 P2 ... Pn
                                       | P : T | { X : U } P | _
Computational context       Γ    ::= ^ | Γ, x : T

module Kind : sig ... end

External computation-level kinds.

module Typ : sig ... end

External computation-level types.

module Expression : sig ... end

External computation-level expressions.

module Pattern : sig ... end

External computation-level patterns.

module Copattern : sig ... end

External computation-level copatterns.

module Context : sig ... end

External computation-level contexts.