Module Comp.Typ

External computation-level types.

type t =

| Constant of {

location : Common.Location.t;

identifier : Common.Qualified_identifier.t;

operator : Common.Operator.t;

quoted : Stdlib.Bool.t;

variant : [ `Inductive | `Stratified | `Coinductive | `Abbreviation ];

}
(*
Constant { identifier = "c"; } is the computation-level type constant "c".

If variant = `Inductive, then this constant was defined in an inductive type declaration.

If variant = `Stratified, then this constant was defined in a stratified type declaration.

If variant = `Coinductive, then this constant was defined in a coinductive type declaration.

If variant = `Abbreviation, then this constant was defined in a computation type abbreviation declaration.

*)

| Arrow of {

location : Common.Location.t;

domain : t;

range : t;

orientation : [ `Forward | `Backward ];

}
(*
Arrow { domain; range; _ } is the computational type domain -> range.
*)

| Cross of {

location : Common.Location.t;

types : t Support.List2.t;

}
(*
Cross { typs = [t1; t2; ...; tn]; _ } is the type of tuple t1 * t2 * ... * tn.
*)