Module Synext.LF

The representation of LF kinds, types, and terms after parsing and data-dependent disambiguation.

ASTs constructed with the constructors in this module are not necessarily in normal form.

These types are only intended to be used in the definition of LF type-level or term-level constants. This is a weak, representational function space without case analysis or recursion.

These types are suited for pretty-printing and elaboration to the internal syntax. Note that this is a named representation.

The metavariable:

x ranges over variables

c ranges over term-level constants

a ranges over type-level constants

LF kinds   K    ::= type | Πx:A.K | A → K
LF types   A, B ::= a | Πx:A.B | A → B | A M1 M2 ... Mn
LF terms   M, N ::= c | x | λx:A.M | M N1 N2 ... Nn | M:A | _

The representation of LF kinds, types, and terms after parsing and data-dependent disambiguation.

ASTs constructed with the constructors in this module are not necessarily in normal form.

These types are only intended to be used in the definition of LF type-level or term-level constants. This is a weak, representational function space without case analysis or recursion.

These types are suited for pretty-printing and elaboration to the internal syntax. Note that this is a named representation.

The metavariable:

x ranges over variables

c ranges over term-level constants

a ranges over type-level constants

LF kinds   K    ::= type | Πx:A.K | A → K
LF types   A, B ::= a | Πx:A.B | A → B | A M1 M2 ... Mn
LF terms   M, N ::= c | x | λx:A.M | M N1 N2 ... Nn | M:A | _

module Kind : sig ... end

External LF kinds.

module Typ : sig ... end

External LF types.

module Term : sig ... end

External LF terms.