Module LF.Term

External LF terms.

type t =

| Variable of {

location : Common.Location.t;

identifier : Common.Identifier.t;

}
(*
Variable { identifier = "x"; _ } is the term-level variable with name "x", which is either a bound variable having a lambda binder, or a free variable having no such corresponding binder.
*)

| Constant of {

location : Common.Location.t;

identifier : Common.Qualified_identifier.t;

operator : Common.Operator.t;

quoted : Stdlib.Bool.t;

}
(*
Constant { identifier = "c"; _ } is the term-level constant with qualified identifier "c".
*)

| Application of {

location : Common.Location.t;

applicand : t;

arguments : t Support.List1.t;

}
(*
Application { applicand; arguments; _ } is the term-level application of applicand with arguments.

If applicand = Term.Constant { operator; _ } and Operator.is_infix operator, then List.length arguments = 2.

If applicand = Term.Constant { operator; _ } and Operator.is_postfix operator, then List.length arguments = 1.

*)

| Abstraction of {

location : Common.Location.t;

parameter_identifier : Common.Identifier.t Support.Option.t;

parameter_type : Typ.t Support.Option.t;

body : t;

}
(*

Abstraction { parameter_identifier = Option.Some "x"; body; _ } is the term \x. body. The variable "x" ranges over LF terms.

Abstraction { parameter_identifier = Option.None; body; _ } is the term \_. body.

*)

| Wildcard of {

location : Common.Location.t;

}
(*
Wildcard { _ } is the omission _ of a fresh term-level variable.
*)

| TypeAnnotated of {

location : Common.Location.t;

term : t;

typ : Typ.t;

}
(*
TypeAnnotated { term = u; typ = t; _ } is the term u : t.
*)