Module Synext.Meta

The representation of meta-types, meta-objects, meta-object patterns, meta-substitutions and meta-contexts after parsing and data-dependent disambiguation.

ASTs constructed with the constructors in this module are not necessarily in normal form.

These types are suited for pretty-printing and elaboration to the internal syntax. Note that this is a named representation.

The metavariable:

X ranges over meta-level variables

g ranges over context schemas

Meta-types           U ::= g | [Ψ ⊢ A] | #[Ψ ⊢ A] | $[Ψ ⊢ Ψ] | $[Ψ ⊢# Ψ]
Meta-objects         C ::= [Ψ] | [Ψ ⊢ M] | $[Ψ ⊢ σ] | $[Ψ ⊢# σ]
Meta-substitutions   θ ::= ^ | θ, C/X
Meta-contexts        Δ ::= ^ | Δ, X:U
Schemas              G ::= g | G + G | some [x1:A1, x2:A2, ..., xn:An] block (y1:B1, y2:B2, ..., ym:Bm)

Meta-object patterns   Cp ::= [Ψp] | [Ψp ⊢ Mp] | $[Ψp ⊢ σp] | $[Ψp ⊢# σp]

The representation of meta-types, meta-objects, meta-object patterns, meta-substitutions and meta-contexts after parsing and data-dependent disambiguation.

ASTs constructed with the constructors in this module are not necessarily in normal form.

These types are suited for pretty-printing and elaboration to the internal syntax. Note that this is a named representation.

The metavariable:

X ranges over meta-level variables

g ranges over context schemas

Meta-types           U ::= g | [Ψ ⊢ A] | #[Ψ ⊢ A] | $[Ψ ⊢ Ψ] | $[Ψ ⊢# Ψ]
Meta-objects         C ::= [Ψ] | [Ψ ⊢ M] | $[Ψ ⊢ σ] | $[Ψ ⊢# σ]
Meta-substitutions   θ ::= ^ | θ, C/X
Meta-contexts        Δ ::= ^ | Δ, X:U
Schemas              G ::= g | G + G | some [x1:A1, x2:A2, ..., xn:An] block (y1:B1, y2:B2, ..., ym:Bm)

Meta-object patterns   Cp ::= [Ψp] | [Ψp ⊢ Mp] | $[Ψp ⊢ σp] | $[Ψp ⊢# σp]

module Typ : sig ... end

External meta-types.

module Object : sig ... end

External meta-objects.

module Pattern : sig ... end

External meta-object patterns.

module Context : sig ... end

External meta-contexts.

module Schema : sig ... end

External context schemas.