Module Meta.Schema

External context schemas.

type t =

| Constant of {

location : Syncom.Location.t;

identifier : Syncom.Qualified_identifier.t;

}
(*
Constant { identifier = "ctx"; _ } is the schema having identifier "ctx" declared elsewhere in the signature.

A tuple term has a block type t matching against this schema if t matches against the schema referred to as `identifier'.
*)

| Alternation of {

location : Syncom.Location.t;

schemas : t Support.List2.t;

}
(*
Alternation { schemas = [g1; g2; ...; gn]; _ } is the schema g1 + g2 + ... + gn.

A tuple term has a block type t matching against this schema if t matches against at least one of g1, g2, ..., gn.
*)