Module Synint.Comp

type unbox_modifier = [

| `Strengthened

]

type case_pragma =

| PragmaCase

| PragmaNotCase

type context_case =

| EmptyContext of Syncom.Location.t

| ExtendedBy of Syncom.Location.t * int

type case_label =

| Lf_constant of Syncom.Location.t * Syncom.Name.t * Syncom.Id.cid_term

| Comp_constant of Syncom.Location.t * Syncom.Name.t * Syncom.Id.cid_comp_const

| BVarCase of Syncom.Location.t

| ContextCase of context_case

| PVarCase of Syncom.Location.t * int * int option

type 'a generic_order =

| Arg of 'a

| Lex of 'a generic_order list

| Simul of 'a generic_order list

type 'a generic_suffices_typ = [

| `exact of 'a

| `infer of Syncom.Location.t

]

Type specified in an interactive use of `suffices`

val map_suffices_typ : ('a -> 'b) -> 'a0 generic_suffices_typ -> 'b0 generic_suffices_typ

val map_order : ('a -> 'b) -> 'a0 generic_order -> 'b0 generic_order

type kind =

| Ctype of Syncom.Location.t

| PiKind of Syncom.Location.t * LF.ctyp_decl * kind

type meta_typ = LF.ctyp

type meta_obj = Syncom.Location.t * LF.mfront

type meta_spine =

| MetaNil

| MetaApp of meta_obj * meta_typ * meta_spine * Syncom.Plicity.t

type typ =

| TypBase of Syncom.Location.t * Syncom.Id.cid_comp_typ * meta_spine

| TypCobase of Syncom.Location.t * Syncom.Id.cid_comp_cotyp * meta_spine

| TypDef of Syncom.Location.t * Syncom.Id.cid_comp_typ * meta_spine

| TypBox of Syncom.Location.t * meta_typ

| TypArr of Syncom.Location.t * typ * typ

| TypCross of Syncom.Location.t * typ Support.List2.t

| TypPiBox of Syncom.Location.t * LF.ctyp_decl * typ

| TypClo of typ * LF.msub

| TypInd of typ

type suffices_typ = typ generic_suffices_typ

val loc_of_typ : typ -> Syncom.Location.t

val loc_of_suffices_typ : suffices_typ -> Syncom.Location.t

type ih_arg =

| M of meta_obj * meta_typ

| V of Syncom.Id.offset

| E

| DC

type wf_tag = bool

type ctyp_decl =

| CTypDecl of Syncom.Name.t * typ * wf_tag

| CTypDeclOpt of Syncom.Name.t

type ih_decl =

| WfRec of Syncom.Name.t * ih_arg list * typ

val rename_ctyp_decl : (Syncom.Name.t -> Syncom.Name.t) -> ctyp_decl -> ctyp_decl

type gctx = ctyp_decl LF.ctx

type ihctx = ih_decl LF.ctx

and exp =

| Fn of Syncom.Location.t * Syncom.Name.t * exp

| Fun of Syncom.Location.t * fun_branches

| MLam of Syncom.Location.t * Syncom.Name.t * exp * Syncom.Plicity.t

| Tuple of Syncom.Location.t * exp Support.List2.t

| LetTuple of Syncom.Location.t * exp * Syncom.Name.t Support.List2.t * exp

| Let of Syncom.Location.t * exp * Syncom.Name.t * exp

| Box of Syncom.Location.t * meta_obj * meta_typ

| Case of Syncom.Location.t * case_pragma * exp * branch list

| Impossible of Syncom.Location.t * exp

| Hole of Syncom.Location.t * Syncom.HoleId.t * Syncom.HoleId.name

| Var of Syncom.Location.t * Syncom.Id.offset

| DataConst of Syncom.Location.t * Syncom.Id.cid_comp_const

| Obs of Syncom.Location.t * exp * LF.msub * Syncom.Id.cid_comp_dest

| Const of Syncom.Location.t * Syncom.Id.cid_prog

| Apply of Syncom.Location.t * exp * exp

| MApp of Syncom.Location.t * exp * meta_obj * meta_typ * Syncom.Plicity.t

| AnnBox of Syncom.Location.t * meta_obj * meta_typ

Normal computational terms

and pattern =

| PatMetaObj of Syncom.Location.t * meta_obj

| PatConst of Syncom.Location.t * Syncom.Id.cid_comp_const * pattern_spine

| PatFVar of Syncom.Location.t * Syncom.Name.t

| PatVar of Syncom.Location.t * Syncom.Id.offset

| PatTuple of Syncom.Location.t * pattern Support.List2.t

| PatAnn of Syncom.Location.t * pattern * typ * Syncom.Plicity.t

and pattern_spine =

| PatNil

| PatApp of Syncom.Location.t * pattern * pattern_spine

| PatObs of Syncom.Location.t * Syncom.Id.cid_comp_dest * LF.msub * pattern_spine

and branch =

| Branch of Syncom.Location.t * LF.mctx * LF.mctx * gctx * pattern * LF.msub * exp

and fun_branches =

| NilFBranch of Syncom.Location.t

| ConsFBranch of Syncom.Location.t * LF.mctx * gctx * pattern_spine * exp * fun_branches

type tclo = typ * LF.msub

type order = int generic_order

type 'order total_dec_kind = [

| `inductive of 'order

| `not_recursive

| `trust

| `partial

]

val map_total_dec_kind : ('o2 -> 'o20) -> 'o21 total_dec_kind -> 'o22 total_dec_kind

val option_of_total_dec_kind : [> `inductive of 'a ] -> 'b option

val apply_many : exp -> exp list -> exp

Applies a spine of checkable terms to a synthesizable term, from left to right.

val loc_of_exp : exp -> Syncom.Location.t

type total_dec = {

name : Syncom.Name.t;

tau : typ;

order : order total_dec_kind;

}

val make_total_dec : Syncom.Name.t -> typ -> order total_dec_kind -> total_dec

val is_meta_obj : exp -> meta_obj option

Decides whether this synthesizable expression is actually an annotated box.

val head_of_meta_obj : meta_obj -> (LF.dctx_hat * LF.head) option

val itermToClObj : LF.iterm -> LF.clobj

val metaObjToMFront : ('a * 'b) -> 'c

val head_of_application : exp -> exp

Finds the head of an application. Chases meta-applications and computation applications.

val strip_pattern : pattern -> pattern

Removes all type annotations from a pattern.

val strip_pattern_spine : pattern_spine -> pattern_spine

type defer_kind = [

| `subgoal

| `theorem

]

type invoke_kind = [

| `ih

| `lemma

]

type split_kind = [

| `split

| `invert

| `impossible

]

type level = [

| `meta

| `comp

]

type hypotheses = {

cD : LF.mctx;

cG : gctx;

cIH : ihctx;

}

val no_hypotheses : hypotheses

type meta_branch_label = [

| `ctor of Syncom.Id.cid_term

| `pvar of int option

| `bvar

]

module SubgoalPath : sig ... end

type proof =

| Incomplete of Syncom.Location.t * proof_state

| Command of command * proof

| Directive of directive

and command =

| By of exp * Syncom.Name.t * typ

| Unbox of exp * Syncom.Name.t * LF.ctyp * unbox_modifier option

and proof_state = {

context : hypotheses;

label : SubgoalPath.builder;

goal : tclo;

solution : proof option Stdlib.ref;

}

and directive =

| Intros of hypothetical

| Solve of exp

| ImpossibleSplit of exp

| Suffices of exp * suffices_arg list

| MetaSplit of exp * typ * meta_branch list

| CompSplit of exp * typ * comp_branch list

| ContextSplit of exp * typ * context_branch list

and suffices_arg = Syncom.Location.t * typ * proof

and context_branch = context_case split_branch

and meta_branch = meta_branch_label split_branch

and comp_branch = Syncom.Id.cid_comp_const split_branch

and 'b split_branch =

| SplitBranch of 'b * gctx * pattern * LF.msub * hypothetical

A general branch of a case analysis.

and hypothetical =

| Hypothetical of Syncom.Location.t * hypotheses * proof

A hypothetical derivation lists meta-hypotheses and hypotheses, then proceeds with a proof.

type open_subgoal = Syncom.Location.t * Syncom.Id.cid_prog * proof_state

An open subgoal is a proof state together with a reference ot the theorem in which it occurs.

val make_proof_state : SubgoalPath.builder -> tclo -> proof_state

Generates a unsolved subgoal with the given goal in an empty context, with no label.

val incomplete_proof : Syncom.Location.t -> proof_state -> proof

Smart constructor for an unfinished proof ending.

val intros : hypotheses -> proof -> proof

Smart constructor for the intros directive.

val suffices : exp -> suffices_arg list -> proof

val proof_cons : command -> proof -> proof

val solve : exp -> proof

val context_split : exp -> typ -> context_branch list -> proof

val context_branch : context_case -> (gctx * pattern) -> LF.msub -> hypotheses -> proof -> context_branch

val meta_split : exp -> typ -> meta_branch list -> proof

val impossible_split : exp -> proof

val meta_branch : meta_branch_label -> (gctx * pattern) -> LF.msub -> hypotheses -> proof -> meta_branch

val comp_split : exp -> typ -> comp_branch list -> proof

val comp_branch : Syncom.Id.cid_comp_const -> (gctx * pattern) -> LF.msub -> hypotheses -> proof -> comp_branch

val prepend_commands : command list -> proof -> proof

Gives a more convenient way of writing complex proofs by using list syntax.

val name_of_ctyp_decl : ctyp_decl -> Syncom.Name.t

val metavariable_of_exp : exp -> (Syncom.Id.offset * Syncom.Id.offset option) option

Decides whether the computational term is actually a variable index object. See `LF.variable_of_mfront`.

val variable_of_exp : exp -> Syncom.Id.offset option

type thm =

| Proof of proof

| Program of exp

type env =

| Empty

| Cons of value * env

and value =

| FnValue of Syncom.Name.t * exp * LF.msub * env

| FunValue of fun_branches_value

| ThmValue of Syncom.Id.cid_prog * thm * LF.msub * env

| MLamValue of Syncom.Name.t * exp * LF.msub * env

| CtxValue of Syncom.Name.t * exp * LF.msub * env

| BoxValue of meta_obj

| ConstValue of Syncom.Id.cid_prog

| DataValue of Syncom.Id.cid_comp_const * data_spine

| TupleValue of value Support.List2.t

and data_spine =

| DataNil

| DataApp of value * data_spine

and fun_branches_value =

| NilValBranch

| ConsValBranch of pattern_spine * exp * LF.msub * env * fun_branches_value