Module Synprs

Parser Syntax Definition

The syntax for Beluga signatures after context-free parsing.

OCaml constructor names prefixed with "Raw" require data-dependent disambiguation or reduction during the elaboration to the external syntax.

Parser LF Syntax

module LF : sig ... end

The intermediate representation of LF kinds, types and terms to delay the handling of data-dependent aspects of the grammar.

Parser Contextual LF Syntax

module CLF : sig ... end

The intermediate representation of contextual LF types, terms and patterns to delay the handling of data-dependent aspects of the grammar.

Parser Meta-Level Syntax

module Meta : sig ... end

The intermediate representation of meta-types, meta-objects and meta-patterns to delay the handling of data-dependent aspects of the grammar.

Parser Computation-Level Syntax

module Comp : sig ... end

The intermediate representation of computation-level kinds, types, expressions and patterns to delay the handling of data-dependent aspects of the grammar.

Parser Harpoon Syntax

module Harpoon : sig ... end

The intermediate representation of Harpoon proof scripts and REPL commands to delay the handling of data-dependent aspects of the grammar.

Parser Signature Syntax

module Signature : sig ... end

The intermediate representation of Beluga signatures to delay the handling of data-dependent aspects of the grammar.

Type Aliases

LF

type lf_object = LF.Object.t

Contextual LF

type clf_object = CLF.Object.t

type clf_context_object = CLF.Context_object.t

Meta-Level

type meta_thing = Meta.Thing.t

type meta_context_object = Meta.Context_object.t

type schema_object = Meta.Schema_object.t

Computation-Level

type comp_sort_object = Comp.Sort_object.t

type comp_expression_object = Comp.Expression_object.t

type comp_pattern_object = Comp.Pattern_object.t

type comp_context_object = Comp.Context_object.t

Harpoon

type harpoon_proof = Harpoon.Proof.t

type harpoon_command = Harpoon.Command.t

type harpoon_directive = Harpoon.Directive.t

type harpoon_split_branch = Harpoon.Split_branch.t

type harpoon_split_branch_label = Harpoon.Split_branch.Label.t

type harpoon_suffices_branch = Harpoon.Suffices_branch.t

type harpoon_hypothetical = Harpoon.Hypothetical.t

type harpoon_repl_command = Harpoon.Repl.Command.t

Signature

type signature_pragma = Signature.Pragma.t

type signature_global_pragma = Signature.Global_pragma.t

type signature_totality_declaration = Signature.Totality.Declaration.t

type 'argument signature_totality_order = 'argument Signature.Totality.Order.t

type signature_declaration = Signature.Declaration.t

type signature_entry = Signature.Entry.t

type signature = Signature.t

Common

include module type of struct include Common end

Operators

module Associativity = Common.Associativity

module Fixity = Common.Fixity

module Operator = Common.Operator

Parameter Annotations

module Plicity = Common.Plicity

module Inductivity = Common.Inductivity

Pretty-Printing

module Parenthesizer = Common.Parenthesizer

Identifiers

module Identifier = Common.Identifier

module Qualified_identifier = Common.Qualified_identifier

Utilities

module Error = Common.Error

module Position = Common.Position

module Location = Common.Location

Locations

LF

val location_of_lf_object : Synprs__.Synprs_definition.LF.Object.t -> Beluga_syntax.Location.t

val set_location_of_lf_object : Beluga_syntax.Location.t -> Synprs__.Synprs_definition.LF.Object.t -> Synprs__.Synprs_definition.LF.Object.t

Contextual LF

val location_of_clf_object : Synprs__.Synprs_definition.CLF.Object.t -> Beluga_syntax.Location.t

val set_location_of_clf_object : Beluga_syntax.Location.t -> Synprs__.Synprs_definition.CLF.Object.t -> Synprs__.Synprs_definition.CLF.Object.t

val location_of_clf_context_object : Synprs__.Synprs_definition.CLF.Context_object.t -> Beluga_syntax.Location.t

Meta-Level

val location_of_meta_thing : Synprs__.Synprs_definition.Meta.Thing.t -> Beluga_syntax.Location.t

val location_of_schema_object : Synprs__.Synprs_definition.Meta.Schema_object.t -> Beluga_syntax.Location.t

val location_of_meta_context_object : Synprs__.Synprs_definition.Meta.Context_object.t -> Beluga_syntax.Location.t

Computation-Level

val location_of_comp_sort_object : Synprs__.Synprs_definition.Comp.Sort_object.t -> Beluga_syntax.Location.t

val set_location_of_comp_sort_object : Beluga_syntax.Location.t -> Synprs__.Synprs_definition.Comp.Sort_object.t -> Synprs__.Synprs_definition.Comp.Sort_object.t

val location_of_comp_expression_object : Synprs__.Synprs_definition.Comp.Expression_object.t -> Beluga_syntax.Location.t

val set_location_of_comp_expression_object : Beluga_syntax.Location.t -> Synprs__.Synprs_definition.Comp.Expression_object.t -> Synprs__.Synprs_definition.Comp.Expression_object.t

val location_of_comp_pattern_object : Synprs__.Synprs_definition.Comp.Pattern_object.t -> Beluga_syntax.Location.t

val set_location_of_comp_pattern_object : Beluga_syntax.Location.t -> Synprs__.Synprs_definition.Comp.Pattern_object.t -> Synprs__.Synprs_definition.Comp.Pattern_object.t

val location_of_comp_context_object : Synprs__.Synprs_definition.Comp.Context_object.t -> Beluga_syntax.Location.t

Harpoon

val location_of_harpoon_proof : Synprs__.Synprs_definition.Harpoon.Proof.t -> Beluga_syntax.Location.t

val location_of_harpoon_command : Synprs__.Synprs_definition.Harpoon.Command.t -> Beluga_syntax.Location.t

val location_of_harpoon_directive : Synprs__.Synprs_definition.Harpoon.Directive.t -> Beluga_syntax.Location.t

val location_of_harpoon_split_branch : Synprs__.Synprs_definition.Harpoon.Split_branch.t -> Beluga_syntax.Location.t

val location_of_harpoon_split_branch_label : Synprs__.Synprs_definition.Harpoon.Split_branch.Label.t -> Beluga_syntax.Location.t

val location_of_harpoon_suffices_branch : Synprs__.Synprs_definition.Harpoon.Suffices_branch.t -> Beluga_syntax.Location.t

val location_of_harpoon_hypothetical : Synprs__.Synprs_definition.Harpoon.Hypothetical.t -> Beluga_syntax.Location.t

val location_of_harpoon_repl_command : Synprs__.Synprs_definition.Harpoon.Repl.Command.t -> Beluga_syntax.Location.t

Signature

val location_of_signature_pragma : Synprs__.Synprs_definition.Signature.Pragma.t -> Beluga_syntax.Location.t

val location_of_signature_global_pragma : Synprs__.Synprs_definition.Signature.Global_pragma.t -> Beluga_syntax.Location.t

val location_of_signature_totality_declaration : Synprs__.Synprs_definition.Signature.Totality.Declaration.t -> Beluga_syntax.Location.t

val location_of_signature_totality_order : 'a Synprs__.Synprs_definition.Signature.Totality.Order.t -> Beluga_syntax.Location.t

val location_of_signature_declaration : Synprs__.Synprs_definition.Signature.Declaration.t -> Beluga_syntax.Location.t