Module Synprs.CLF

The intermediate representation of contextual LF types, terms and patterns to delay the handling of data-dependent aspects of the grammar.

This is LF augmented with substitutions, contexts and blocks.

The intermediate representation of contextual LF types, terms and patterns to delay the handling of data-dependent aspects of the grammar.

This is LF augmented with substitutions, contexts and blocks.

module Object : sig ... end

Contextual LF types, terms and patterns blurred together.

module Context_object : sig ... end

Contextual LF substitutions and contexts blurred together.