Module Synprs.Comp

The intermediate representation of computation-level kinds, types, expressions and patterns to delay the handling of data-dependent aspects of the grammar.

The intermediate representation of computation-level kinds, types, expressions and patterns to delay the handling of data-dependent aspects of the grammar.

module Sort_object : sig ... end

Computational kinds and types blurred together.

module Expression_object : sig ... end

module Pattern_object : sig ... end

module Copattern_object : sig ... end

module Context_object : sig ... end