Module Synprs.Harpoon

The intermediate representation of Harpoon proof scripts and REPL commands to delay the handling of data-dependent aspects of the grammar.

The intermediate representation of Harpoon proof scripts and REPL commands to delay the handling of data-dependent aspects of the grammar.

module Proof : sig ... end

module Command : sig ... end

module Directive : sig ... end

module Split_branch : sig ... end

module Suffices_branch : sig ... end

module Hypothetical : sig ... end

module Repl : sig ... end