Module Beluga.ConvSigma

author Brigitte Pientka

type t

val fmt_ppr_conv_list : Stdlib.Format.formatter -> t -> unit

val map : t -> Beluga_syntax.Id.offset -> Beluga_syntax.Id.offset

Translates a bound variable index according to the sigma conversion.

val flattenDCtx : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.dctx * t

val gen_proj_sub : t -> Beluga_syntax.Synint.LF.sub

val gen_tup_sub : t -> Beluga_syntax.Synint.LF.sub

val gen_proj_sub' : t -> Beluga_syntax.Synint.LF.sub

val gen_tup_sub' : t -> Beluga_syntax.Synint.LF.sub

val etaExpandMMVstr : Beluga_syntax.Location.t -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.tclo -> Beluga_syntax.Plicity.t -> Beluga_syntax.Name.t option -> Beluga_syntax.Name.t list -> Beluga_syntax.Synint.LF.normal

Constructs a unification variable for the given tclo, strengthening its type.

val etaExpandMPVstr : Beluga_syntax.Location.t -> Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.tclo -> Beluga_syntax.Plicity.t -> Beluga_syntax.Name.t option -> Beluga_syntax.Name.t list -> Beluga_syntax.Synint.LF.normal

Constructs a unification parameter variable for a given tclo, strengthening its type.

val gen_flattening : Beluga_syntax.Synint.LF.mctx -> Beluga_syntax.Synint.LF.dctx -> Beluga_syntax.Synint.LF.dctx * Beluga_syntax.Synint.LF.sub Stdlib.Lazy.t * Beluga_syntax.Synint.LF.sub Stdlib.Lazy.t

gen_flattening cD cPsi = (cPhi, lazy s_proj, lazy s_tup) Generates a flattened LF context cPhi in which all blocks present in cPsi have been decomposed. Packing and unpacking substitutions s_proj and s_tup are lazily computed to mediate between the contexts. Specifically: if cD |- cPsi ctx then cD |- cPhi ctx such that cD ; cPsi |- s_proj : cPhi cD ; cPhi |- s_tup : cPsi

The naming of s_proj and s_tup is based on what kind of terms these substitutions *contain*; e.g. s_proj contains projections, so it takes us from the block context to the flat context.