Module Synprs.LF

The intermediate representation of LF kinds, types and terms to delay the handling of data-dependent aspects of the grammar.

The parsers associated with these types are only intended to be used in the definition of LF type-level or term-level constants. This is a weak, representational function space without case analysis or recursion.

The intermediate representation of LF kinds, types and terms to delay the handling of data-dependent aspects of the grammar.

The parsers associated with these types are only intended to be used in the definition of LF type-level or term-level constants. This is a weak, representational function space without case analysis or recursion.

module Object : sig ... end

LF kinds, types and terms blurred together.